% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_function.R
\name{agricolae_correlation}
\alias{agricolae_correlation}
\title{Title}
\usage{
agricolae_correlation(
  x,
  y = NULL,
  method = c("pearson", "kendall", "spearman", "lin"),
  alternative = "two.sided"
)
}
\arguments{
\item{x}{wait_for_add}

\item{y}{wait_for_add}

\item{method}{wait_for_add}

\item{alternative}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
