% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_gason.R
\name{build_gson}
\alias{build_gson}
\title{Title}
\usage{
build_gson(keyType = "ko", KEGG_Type = "KEGG", species = "hsa")
}
\arguments{
\item{keyType}{keyType}

\item{KEGG_Type}{keyType}

\item{species}{keyType}
}
\value{
gson object
}
\description{
Title
}
\examples{
# xx
}
