% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_diff_analysis.R
\name{chectCores}
\alias{chectCores}
\title{Title}
\usage{
chectCores(all.tests = FALSE, logical = TRUE)
}
\arguments{
\item{all.tests}{wait_for_add}

\item{logical}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
