% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_Mi.R
\name{cmi}
\alias{cmi}
\title{Conditional mutual information}
\usage{
cmi(genelist1, genelist2, genelist3)
}
\arguments{
\item{genelist1}{genelist1}

\item{genelist2}{genelist2}

\item{genelist3}{genelist3}
}
\description{
Conditional mutual information
}
