% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorRcpp.R
\name{CorRcpp}
\alias{CorRcpp}
\title{Fast correlation analysis}
\usage{
CorRcpp(x = x, y = NULL, type = c("pearson", "spearman"))
}
\arguments{
\item{x}{Dataframe.}

\item{y}{Dataframe.}

\item{type}{A character string. Methods include pearson (default), spearman.}
}
\value{
list
}
\description{
Fast correlation analysis
}
\examples{
data_df1 <- data.frame(
  A = c(11, 23, 3, 45, 5),
  B = c(62, 7, 82, 9, 10),
  C = c(11, 54, 13, 42, 15),
  D = c(16, 42, 18, 23, 20),
  E = c(12, 54, 23, 33, 42)
)

data_df2 <- data.frame(
  Tom = c(11, 27, 23, 29, 30),
  Anny = c(31, 16, 33, 98, 35),
  Jerry = c(36, 37, 45, 39, 40),
  Cat = c(41, 21, 43, 44, 45),
  White = c(46, 77, 48, 94, 12)
)
re <- CorRcpp(data_df1,data_df2,type = 'spearman')
re$R_matrix # correlation value
re$P_matrix # pvalue
}
\author{
Qiusheng Wu
}
