% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_function.R
\name{correl}
\alias{correl}
\title{Title}
\usage{
correl(x, y, method = "pearson", alternative = "two.sided")
}
\arguments{
\item{x}{wait_for_add}

\item{y}{wait_for_add}

\item{method}{wait_for_add}

\item{alternative}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
