% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_data_utlis.R, R/EMP_estimate_sample.R
\name{cosine}
\alias{cosine}
\title{Title}
\usage{
cosine(vector1, vector2)

cosine(vector1, vector2)
}
\arguments{
\item{vector1}{wait_for_add}

\item{vector2}{wait_for_add}
}
\value{
xx object

xx object
}
\description{
Title

Title
}
\examples{
# add example
# add example
}
