% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Title}
\usage{
create_progress_bar(name = "none", ...)
}
\arguments{
\item{name}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
