% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_convert.R
\name{.EMP_feature_convert_cpd}
\alias{.EMP_feature_convert_cpd}
\title{Title}
\usage{
.EMP_feature_convert_cpd(EMPT, method = "mean", from, to)
}
\arguments{
\item{EMPT}{EMPT}

\item{method}{method}

\item{from}{from}

\item{to}{to}
}
\value{
EMPT object
}
\description{
Title
}
\examples{
#
}
