% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_convert.R
\name{.EMP_feature_convert_gene}
\alias{.EMP_feature_convert_gene}
\title{Title}
\usage{
.EMP_feature_convert_gene(
  EMPT,
  method = "mean",
  from,
  to,
  species = "Human",
  OrgDb = NULL
)
}
\arguments{
\item{from}{from}

\item{to}{to}

\item{species}{species}

\item{OrgDb}{OrgDb}

\item{feature}{features}
}
\value{
data.frame
}
\description{
Title
}
\examples{
#
}
