% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_support.R
\name{.pattern_Dectect_multi}
\alias{.pattern_Dectect_multi}
\title{Title}
\usage{
.pattern_Dectect_multi(pattern_ref, pattern, exact = F)
}
\arguments{
\item{pattern_ref}{Object in EMPT format.}

\item{pattern}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{exact}{A character string.Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq", "chord", "hellinger", "aitchison", or "robust.aitchison".}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
