% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{EMP-class}
\alias{EMP-class}
\title{Class "EMP"
This class represents the ...}
\description{
Class "EMP"
This class represents the ...
}
\section{Slots}{

\describe{
\item{\code{ExperimentList}}{ExperimentList.}

\item{\code{deposit}}{deposit.}

\item{\code{plot_deposit}}{plot_deposit.}

\item{\code{message_info}}{message_info}

\item{\code{history}}{history.}

\item{\code{method}}{method.}

\item{\code{palette}}{palette.}

\item{\code{plot_category}}{plot_category.}

\item{\code{plot_specific}}{plot_specific.}

\item{\code{plot_info}}{plot_info.}

\item{\code{info}}{info.}
}}

\author{
Bingdong Liu
}
\keyword{classes}
