% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_adjust_abudance.R
\name{EMP_adjust_abudance}
\alias{EMP_adjust_abudance}
\title{Title}
\usage{
EMP_adjust_abudance(
  x,
  experiment,
  method = "combat_seq",
  use_cached = T,
  .factor_unwanted,
  .factor_of_interest,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{method}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{.factor_unwanted}{wait_for_add}

\item{.factor_of_interest}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
