% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparetion_adjust_abundance.R
\name{EMP_adjust_abundance}
\alias{EMP_adjust_abundance}
\title{Adjust experssion or abundance for unexpected bias or batch effect}
\usage{
EMP_adjust_abundance(
  obj,
  experiment,
  method = "combat_seq",
  use_cached = TRUE,
  .factor_unwanted,
  .factor_of_interest,
  action = "add",
  ...
)
}
\arguments{
\item{obj}{Object in EMPT or SummarisedExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Methods include combat_seq (default), combat and limma_remove_batch_effect.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{.factor_unwanted}{A tidy select, e.g. column names without double quotation. c(batch, country) These are the factor that we want to adjust for, including unwanted batcheffect, and unwanted biological effects.}

\item{.factor_of_interest}{A tidy select, e.g. column names without double quotation. c(treatment) These are the factor that we want to preserve.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function tidybulk::adjust_abundance.}
}
\value{
EMPT object
}
\description{
Adjust experssion or abundance for unexpected bias or batch effect
}
\examples{
data(MAE)
MAE |>
  EMP_assay_extract(experiment='geno_ko') |>
  EMP_adjust_abundance(.factor_unwanted = 'Region',.factor_of_interest = 'Group',
                      method = 'combat_seq',action = 'add') 
## combat method 
MAE |>
  EMP_assay_extract(experiment='geno_ko') |>
  EMP_adjust_abundance(.factor_unwanted = 'Region',.factor_of_interest = 'Group',
                      method = 'combat',action = 'add') 
## limma_remove_batch_effect
MAE |>
  EMP_assay_extract(experiment='geno_ko') |>
  EMP_adjust_abundance(.factor_unwanted = 'Region',.factor_of_interest = 'Group',
                      method = 'limma_remove_batch_effect') 
}
