% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_alpha_analysis.R
\name{EMP_alpha_analysis}
\alias{EMP_alpha_analysis}
\title{Title}
\usage{
EMP_alpha_analysis(x, experiment, use_cached = T, action = "add", ...)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
