% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{EMP_assay_extract}
\alias{EMP_assay_extract}
\alias{EMP_assay_extract,MultiAssayExperiment-method}
\alias{EMP_assay_extract,EMPT-method}
\title{Title}
\usage{
EMP_assay_extract(obj, ...)

\S4method{EMP_assay_extract}{MultiAssayExperiment}(obj, ...)

\S4method{EMP_assay_extract}{EMPT}(obj, ...)
}
\arguments{
\item{obj}{obj}

\item{...}{...}

\item{MultiAssayExperiment}{MultiAssayExperiment}

\item{EMPT}{xx}
}
\value{
xx object

xx object

xx object
}
\description{
Title

Title

Title
}
\examples{
# xx
# 
#
}
