% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_EMP_barplot.R
\name{EMP_barplot}
\alias{EMP_barplot}
\title{Barplot for EMPT result}
\usage{
EMP_barplot(
  obj,
  method = "wilcox.test",
  estimate_group = NULL,
  compare_group = NULL,
  group_level = "default",
  compare_group_level = "default",
  error_bar = "default",
  label = "p",
  dot_size = 3,
  dot_color = "white",
  bar_alpha = 1,
  bar_color = "black",
  step.increase = 0.1,
  tip.length = 0,
  ref.group = NULL,
  comparisons = NULL,
  hide_point = FALSE,
  paired_group = NULL,
  paired_line = TRUE,
  p.adjust = "fdr",
  ncol = NULL,
  show = "pic",
  seed = 123,
  palette = NULL,
  test.arg = list(),
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()",
  y_break = NULL,
  facet = TRUE,
  ...
)
}
\arguments{
\item{obj}{EMPT object.}

\item{method}{statistics including t.test, wilcox.test, sign.test, emmeans.test, dunn.test, tukey.hsd.}

\item{estimate_group}{The primary grouping variable for estimation and comparison.Serves as the comparison group by default.}

\item{compare_group}{Optional subgroup analysis variable. When specified, the subgroup comparison performs.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{compare_group_level}{A string vector. Set the compare group order in the plot.}

\item{error_bar}{A character string including default and upper.}

\item{label}{A character string including p, p.adj, p.signif, p.adj.signif.}

\item{dot_size}{A numeric. Set the dot size.}

\item{dot_color}{A color character string.Dot color including white, black, ... or "group" for grouping-based colors.}

\item{bar_alpha}{A numeric. Set the bar alpha.}

\item{bar_color}{A color character to set the border of bar. Border color including black, ... or "group" for grouping-based colors.}

\item{step.increase}{A numeric vector with the increase in fraction of total height for every additional comparison to minimize overlap.}

\item{tip.length}{A numeric vector with the fraction of total height that the bar goes down to indicate the precise column}

\item{ref.group}{A character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of 2 values on the x-axis or the 2 integers that correspond to the index of the columns of interest.(default:NULL)}

\item{hide_point}{A boolean. Whether show the dot or not.}

\item{paired_group}{A character string. Variable name corresponding to paired primary or sample.}

\item{paired_line}{A boolean. Whether show the paired line when paired test activated.(defalut:TRUE)}

\item{p.adjust}{A character string. Adjust P-values for Multiple Comparisons inluding fdr, holm, hochberg, hommel, bonferroni, BH, BY. (default:fdr)}

\item{ncol}{An interger. Set the col number in the facet plot.}

\item{show}{A character string include pic (default), html.}

\item{seed}{An interger. Set the random seed for the point layout.}

\item{palette}{A series of character string. Color palette.}

\item{test.arg}{Additional arguments for the test method.}

\item{html_width}{html_width}

\item{html_height}{html_height}

\item{mytheme}{Modify components of a theme according to the \code{\link[ggplot2]{theme}} and \code{\link[ggplot2]{ggtheme}}.}

\item{y_break}{Break point.}

\item{facet}{If TRUE (default), each variable is displayed in a separate panel. If FALSE, all variables are overlaid in a single panel.}

\item{...}{additional parameters, see also \code{\link[ggpubr]{stat_pvalue_manual}}}
}
\description{
Barplot for EMPT result
}
\examples{
data(MAE)
# plot
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group')

# set error bar
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',error_bar = 'upper')

# for subgroup
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='Sex')

# set signif
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='Sex',label = 'p.signif')

# combine the plot
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='Sex',facet = FALSE)

# set the comparision
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='sub_group',
              ref.group = 'A',
              facet = FALSE)
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='sub_group',
              comparisons = list(c('A','B')),
              facet = FALSE)

# for paired test
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'sub_group',
              paired_group = 'patient')

# plot as prism style
if (!requireNamespace("ggprism", quietly=TRUE)) install.packages("ggprism")
library(ggprism)
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='Status',
              palette = c('white','red','black'),error_bar = 'upper',
              method = 'tukey.hsd',facet = FALSE,
              mytheme ='theme_prism()')
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1BG') |> 
  EMP_barplot(estimate_group = 'Group',compare_group='Status',
              bar_alpha = 0,bar_color = 'group',
              palette = c('yellow','red','green'),error_bar = 'upper',
              method = 'tukey.hsd',dot_color = 'group',
              mytheme ='theme_prism()')
}
