% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_alpha_boxplot.R
\name{EMP_boxplot_alpha}
\alias{EMP_boxplot_alpha}
\title{Title}
\usage{
EMP_boxplot_alpha(EMPT, plot_category = 1, seed = 123, ...)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{plot_category}{wait_for_add}

\item{seed}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
