% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_boxplot_assay.R
\name{EMP_boxplot_assay_default}
\alias{EMP_boxplot_assay_default}
\title{Title}
\usage{
EMP_boxplot_assay_default(
  EMPT,
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  ncol = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{EMPT}{Object in EMPT format.}

\item{method}{The name of the statistical test that is applied to the values of the 2 columns (e.g. t.test, wilcox.test etc.).}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{ncol}{An interger. Set the col number in the facet plot.}

\item{show}{A character string include pic (default), html.}

\item{palette}{A character string.Color palette.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
