% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_boxplot_assay.R
\name{EMP_boxplot_raw}
\alias{EMP_boxplot_raw}
\title{Title}
\usage{
EMP_boxplot_raw(
  obj,
  method = "wilcox.test",
  x = "Group",
  y = "abundance",
  fill = "Group",
  facet_wrap = "ID",
  long_data = F,
  html = F,
  html_width = NULL,
  html_height = NULL
)
}
\arguments{
\item{obj}{wait_for_add}

\item{method}{wait_for_add}

\item{x}{wait_for_add}

\item{y}{wait_for_add}

\item{fill}{wait_for_add}

\item{facet_wrap}{wait_for_add}

\item{long_data}{wait_for_add}

\item{html}{wait_for_add}

\item{html_width}{wait_for_add}

\item{html_height}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
