% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_cluster_analysis.R
\name{EMP_cluster_analysis}
\alias{EMP_cluster_analysis}
\title{Title}
\usage{
EMP_cluster_analysis(
  x,
  experiment,
  distance = "bray",
  rowdata = F,
  method = "average",
  h = NULL,
  groupLabels = T,
  cluster = NULL,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{distance}{wait_for_add}

\item{rowdata}{wait_for_add}

\item{method}{wait_for_add}

\item{h}{wait_for_add}

\item{groupLabels}{wait_for_add}

\item{cluster}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
