% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_EMP_extract.R
\name{EMP_coldata_extract}
\alias{EMP_coldata_extract}
\title{Extract coldata}
\usage{
EMP_coldata_extract(
  obj,
  experiment = NULL,
  coldata_to_assay = NULL,
  assay_name = "undefined",
  action = "get"
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{coldata_to_assay}{A series of character strings. The coldata_to_assay used in action = "add". Select which columns in the coldata to transfer into assay.}

\item{assay_name}{A character string. The assay_name used in action = "add".(default:undefined)}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object or table
}
\description{
Extract coldata
}
\examples{
data(MAE)
## Extract the coldata/meta-data/sample-info/patient-info of one existed experiment from MultiAssaayExperiment
MAE |>
  EMP_coldata_extract('taxonomy')

## Extract all coldata
MAE |>
  EMP_coldata_extract(action = 'get')  # when action = get, output is a tibble.

MAE |>
  EMP_coldata_extract(action = 'add')  # when action = add, output is a EMPT object for downstreanm analysis.
}
