% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_data_utlis.R
\name{EMP_collapse}
\alias{EMP_collapse}
\title{Title}
\usage{
EMP_collapse(
  x,
  experiment,
  estimate_group = NULL,
  method = "sum",
  na_string = c("NA", "null", ""),
  collapse_by,
  collapse_sep = " ",
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{method}{wait_for_add}

\item{na_string}{wait_for_add}

\item{collapse_by}{wait_for_add}

\item{collapse_sep}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
