% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_cor_analysis.R
\name{EMP_cor_analysis}
\alias{EMP_cor_analysis}
\title{Title}
\usage{
EMP_cor_analysis(
  EMP,
  select = NULL,
  method = "spearman",
  action = "add",
  use_cached = T,
  ...
)
}
\arguments{
\item{EMP}{wait_for_add}

\item{select}{wait_for_add}

\item{method}{wait_for_add}

\item{action}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
