% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_heatmap.R
\name{EMP_COR_HEAT_}
\alias{EMP_COR_HEAT_}
\title{Title}
\usage{
EMP_COR_HEAT_(
  data,
  meta_data,
  method = "spearman",
  width = 10,
  height = 10,
  cor_output = F,
  cluster_rows = F,
  cluster_cols = F,
  file_name = "cor_plot"
)
}
\arguments{
\item{data}{wait_for_add}

\item{meta_data}{wait_for_add}

\item{method}{wait_for_add}

\item{width}{wait_for_add}

\item{height}{wait_for_add}

\item{cor_output}{wait_for_add}

\item{cluster_rows}{wait_for_add}

\item{cluster_cols}{wait_for_add}

\item{file_name}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
