% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{EMP_curveplot}
\alias{EMP_curveplot}
\alias{EMP_curveplot,EMP_enrich_analysis-method}
\title{Title}
\usage{
EMP_curveplot(obj, ...)

\S4method{EMP_curveplot}{EMP_enrich_analysis}(obj, ...)
}
\arguments{
\item{obj}{obj}

\item{...}{...}
}
\value{
xx object

xx object
}
\description{
Title

Title
}
\examples{
#
#
}
