% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_decostand.R
\name{EMP_decostand}
\alias{EMP_decostand}
\title{Title}
\usage{
EMP_decostand(
  x,
  experiment,
  method,
  bySample = T,
  logbase = 2,
  use_cached = T,
  pseudocount = 1e-07,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{method}{wait_for_add}

\item{bySample}{wait_for_add}

\item{logbase}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{pseudocount}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
