% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_diff_analysis.R
\name{EMP_diff_analysis}
\alias{EMP_diff_analysis}
\title{Title}
\usage{
EMP_diff_analysis(
  x,
  experiment,
  method = "wilcox.test",
  estimate_group = NULL,
  use_cached = T,
  action = "add",
  group_level = NULL,
  core = NULL,
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{method}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{action}{wait_for_add}

\item{group_level}{wait_for_add}

\item{core}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
