% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_dimension.R
\name{EMP_dimension_analysis}
\alias{EMP_dimension_analysis}
\title{Dimension reduction of the abundance or experssion data}
\usage{
EMP_dimension_analysis(
  x,
  experiment,
  method = "pcoa",
  distance = NULL,
  use_cached = TRUE,
  estimate_group = NULL,
  action = "add"
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Methods include pca, pcoa, pls, opls.}

\item{distance}{A character string. The logarithm distance used in method = "pcoa".Detailed in the vegan::vegdist.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{estimate_group}{A character string. Select the group name in the coldata to be considerated.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Dimension reduction of the abundance or experssion data
}
\examples{
data(MAE)
## PCA
MAE |>
  EMP_dimension_analysis(experiment = 'taxonomy',method = 'pca') 
## Pcoa
MAE |>
  EMP_dimension_analysis(experiment = 'taxonomy',method = 'pcoa',distance = 'bray')
## Pls
MAE |>
  EMP_dimension_analysis(experiment = 'untarget_metabol',
                         method = 'pls',estimate_group = 'Group')
MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |> ## Get the kegg compound
  EMP_dimension_analysis(method = 'pls',estimate_group = 'Group')
## OPLS
MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group) |>
  EMP_filter(feature_condition = pvalue < 0.05) |>
  EMP_dimension_analysis(method = 'opls',estimate_group = 'Sex') |>
  EMP_scatterplot(estimate_group='Sex',show='p12html',ellipse=0.6) ## Visualization
}
