% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/EMP_enrich_dotplot.R
\name{EMP_dotplot}
\alias{EMP_dotplot}
\alias{EMP_dotplot,EMP_enrich_analysis-method}
\alias{EMP_dotplot_enrich}
\title{Dotplot for enrichment result}
\usage{
EMP_dotplot(obj, ...)

\S4method{EMP_dotplot}{EMP_enrich_analysis}(obj, ...)

EMP_dotplot_enrich(
  EMPT,
  plot_category = 1,
  seed = 123,
  color = "p.adjust",
  showCategory = 10,
  ...
)
}
\arguments{
\item{obj}{object}

\item{...}{Further parameters passed to enrichplot::dotplot.}

\item{EMPT}{EMPT object}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{color}{A character string. Variable that used to color enriched terms, e.g. 'pvalue', 'p.adjust' or 'qvalue'}

\item{showCategory}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{EMP_enrich_analysis}{EMP_enrich_analysis}
}
\value{
EMP_enrich_analysis_dotplot object

EMP_enrich_analysis_dotplot object

xx object
}
\description{
Dotplot for enrichment result

Title

Dotplot for enrichment result
}
\examples{
# 
#
# add example
}
