% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_enrich_dotplot.R
\name{EMP_dotplot_enrich}
\alias{EMP_dotplot_enrich}
\title{Title}
\usage{
EMP_dotplot_enrich(EMPT, plot_category = 1, seed = 123, ...)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{plot_category}{wait_for_add}

\item{seed}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
