% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_enrich_dotplot.R
\name{EMP_dotplot_enrich_default}
\alias{EMP_dotplot_enrich_default}
\title{Title}
\usage{
EMP_dotplot_enrich_default(EMPT, show = "pic", ...)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{show}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
