% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_enrich_analysis.R
\name{EMP_enrich_analysis}
\alias{EMP_enrich_analysis}
\title{Title}
\usage{
EMP_enrich_analysis(x, ..., action = "add")
}
\arguments{
\item{x}{wait_for_add}

\item{...}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
