% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_enrich_dotplot.R
\name{EMP_enrich_dotplot}
\alias{EMP_enrich_dotplot}
\alias{EMP_enrich_dotplot,EMP_enrich_analysis-method}
\alias{EMP_enrich_dotplot,EMP_multi_diff_enrich-method}
\alias{EMP_enrich_dotplot,EMP_multi_same_enrich-method}
\alias{EMP_dotplot_enrich}
\title{Dotplot for enrichment result}
\usage{
EMP_enrich_dotplot(obj, ...)

\S4method{EMP_enrich_dotplot}{EMP_enrich_analysis}(obj, ...)

\S4method{EMP_enrich_dotplot}{EMP_multi_diff_enrich}(obj, ...)

\S4method{EMP_enrich_dotplot}{EMP_multi_same_enrich}(obj, ...)

EMP_dotplot_enrich(
  obj,
  plot_category = 1,
  seed = 123,
  color = "p.adjust",
  showCategory = 10,
  ...
)
}
\arguments{
\item{obj}{EMPT or EMP object}

\item{...}{Addtional parameters, see also \code{\link[enrichplot]{dotplot}}.}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{color}{A character string. Variable that used to color enriched terms, e.g. 'pvalue', 'p.adjust' or 'qvalue'}

\item{showCategory}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{EMP_enrich_analysis}{EMP_multi_enrich}
}
\value{
Enrichment dotplot
}
\description{
Dotplot for enrichment result
}
\examples{
\dontrun{
data(MAE)
MAE |>
  EMP_GSEA_analysis(experiment = 'geno_ko',method='signal2Noise',
                    estimate_group = 'Group',
                    pvalueCutoff = 0.05,keyType = 'ko') |>
  EMP_enrich_dotplot(color='p.adjust',showCategory=10) 
}
# add example
}
