% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_enrich_analysis.R
\name{EMP_enrich_kegg}
\alias{EMP_enrich_kegg}
\title{Title}
\usage{
EMP_enrich_kegg(
  data,
  mapID,
  metabolite_col = "blue",
  gene_col = "red",
  gradient_by = NULL,
  gradient_col = c("pink", "steelblue")
)
}
\arguments{
\item{data}{wait_for_add}

\item{mapID}{wait_for_add}

\item{metabolite_col}{wait_for_add}

\item{gene_col}{wait_for_add}

\item{gradient_by}{wait_for_add}

\item{gradient_col}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
