% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_enrich_netplot.R
\name{EMP_enrich_netplot}
\alias{EMP_enrich_netplot}
\alias{EMP_enrich_netplot,EMP_enrich_analysis-method}
\alias{EMP_enrich_netplot,EMP_multi_same_enrich-method}
\alias{EMP_enrich_netplot,EMP_multi_diff_enrich-method}
\alias{EMP_netplot_enrich}
\title{Netplot for enrichment result}
\usage{
EMP_enrich_netplot(obj, ...)

\S4method{EMP_enrich_netplot}{EMP_enrich_analysis}(obj, ...)

\S4method{EMP_enrich_netplot}{EMP_multi_same_enrich}(obj, ...)

\S4method{EMP_enrich_netplot}{EMP_multi_diff_enrich}(obj, ...)

EMP_netplot_enrich(obj, plot_category = 1, seed = 123, showCategory = 10, ...)
}
\arguments{
\item{obj}{EMP or EMPT object}

\item{...}{Additional parameters, see also \code{\link[enrichplot]{cnetplot}}}

\item{plot_category}{An interger. More plot style.(under constrution)}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{showCategory}{A number or a vector of terms. If it is a number, the first n terms will be displayed. If it is a vector of terms, the selected terms will be displayed.}
}
\value{
Enrichment netplot object
}
\description{
Netplot for enrichment result
}
\examples{
\dontrun{
data(MAE)
MAE |>
  EMP_GSEA_analysis(experiment = 'geno_ko',method='signal2Noise',
                    estimate_group = 'Group',
                    pvalueCutoff = 0.05,keyType = 'ko') |>
  EMP_enrich_netplot(showCategory=10) 
}
}
