% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_enrich_analysis.R
\name{EMP_enrich_plot_raw}
\alias{EMP_enrich_plot_raw}
\title{Title}
\usage{
EMP_enrich_plot_raw(
  df,
  type,
  aes = "dot",
  showCategory = 20,
  minGSSize = 1,
  maxGSSize = 500,
  show_data = F,
  use_cache = T
)
}
\arguments{
\item{df}{wait_for_add}

\item{type}{wait_for_add}

\item{aes}{wait_for_add}

\item{showCategory}{wait_for_add}

\item{minGSSize}{wait_for_add}

\item{maxGSSize}{wait_for_add}

\item{show_data}{wait_for_add}

\item{use_cache}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
