% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_data_utlis.R
\name{EMP_estimate_sample}
\alias{EMP_estimate_sample}
\title{Title}
\usage{
EMP_estimate_sample(beta_obj, top_number = NULL)
}
\arguments{
\item{beta_obj}{wait_for_add}

\item{top_number}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
