% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_feature_covert.R
\name{EMP_feature_convert}
\alias{EMP_feature_convert}
\title{Covert featureID of gene experssion or compund abundance}
\usage{
EMP_feature_convert(
  x,
  experiment,
  method = "mean",
  from,
  to,
  species = "none",
  OrgDb = NULL,
  action = "add"
)
}
\arguments{
\item{x}{EMPT or MultiAssayExperiment object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Methods include mean, sum, median, min, max. When multiple annotations appear on features, merge activate.}

\item{from}{A character string. For metabolite include CAS,DTXSID,DTXCID,SID,CID,KEGG,ChEBI,HMDB,Drugbank. For gene include SYMBOL,ENSEMBL,ENTREZID.}

\item{to}{A character string. For metabolite include CAS,DTXSID,DTXCID,SID,CID,KEGG,ChEBI,HMDB,Drugbank. For gene include SYMBOL,ENSEMBL,ENTREZID.}

\item{species}{A character string. Species includ Human,Mouse,Pig,Zebrafish. If converting feature from other species,please use OrgDb.}

\item{OrgDb}{Supported OrgDb listed in 'https://bioconductor.org/packages/release/BiocViews.html#___OrgDb'}

\item{action}{A character string. A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Covert featureID of gene experssion or compund abundance
}
\examples{
#
}
