% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_filter.R
\name{EMP_filter}
\alias{EMP_filter}
\title{Title}
\usage{
EMP_filter(
  obj,
  sample_condition,
  feature_condition,
  filterSample = NULL,
  filterFeature = NULL,
  experiment = NULL,
  show_info = NULL,
  action = "select"
)
}
\arguments{
\item{obj}{wait_for_add}

\item{sample_condition}{wait_for_add}

\item{feature_condition}{wait_for_add}

\item{filterSample}{wait_for_add}

\item{filterFeature}{wait_for_add}

\item{experiment}{wait_for_add}

\item{show_info}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
