% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_tidy_method.R
\name{EMP_filter_deprecated}
\alias{EMP_filter_deprecated}
\title{Title}
\usage{
EMP_filter_deprecated(x, condition, experiment = NULL)
}
\arguments{
\item{x}{wait_for_add}

\item{condition}{wait_for_add}

\item{experiment}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
