% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_filter.R
\name{EMP_filterBysample}
\alias{EMP_filterBysample}
\title{Title}
\usage{
EMP_filterBysample(obj, condition, action = "select")
}
\arguments{
\item{obj}{wait_for_add}

\item{condition}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
