% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_fitline.R
\name{EMP_fitline_plot}
\alias{EMP_fitline_plot}
\alias{EMP_fitline_plot,EMPT-method}
\alias{EMP_fitline_plot,EMP-method}
\alias{EMP_fitline_plot.EMPT}
\alias{EMP_fitline_plot.EMP}
\title{Two variable regression plot}
\usage{
EMP_fitline_plot(obj, ...)

\S4method{EMP_fitline_plot}{EMPT}(obj, ...)

\S4method{EMP_fitline_plot}{EMP}(obj, ...)

EMP_fitline_plot.EMPT(
  obj,
  plot_category = 1,
  seed = 123,
  var_select,
  estimate_group = NULL,
  formula = y ~ poly(x, 1, raw = FALSE),
  se = FALSE,
  show = "pic",
  palette = NULL,
  eq_size = 3,
  mytheme = "theme()",
  html_width = NULL,
  html_height = NULL
)

EMP_fitline_plot.EMP(
  obj,
  plot_category = 1,
  seed = 123,
  var_select,
  select = NULL,
  estimate_group = NULL,
  formula = y ~ poly(x, 1, raw = FALSE),
  se = FALSE,
  show = "pic",
  palette = NULL,
  eq_size = 3,
  mytheme = "theme()",
  html_width = NULL,
  html_height = NULL
)
}
\arguments{
\item{obj}{EMPT or EMP object}

\item{...}{...}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{seed}{An interger. Set the random seed to the plot.}

\item{var_select}{An seris of string. Select two variable to regression.}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{formula}{a formula object. Using aesthetic names x and y instead of original variable names. [y~poly(x,1,raw = FALSE)]. Detailed in ggpmisc::stat_poly_line}

\item{se}{An boolean. Display confidence interval around smooth. [Default:False]}

\item{show}{A character string include pic (default), html.}

\item{palette}{A series of character string. Color palette.}

\item{eq_size}{An number. Equation label size. [Default:3]}

\item{mytheme}{Modify components of a theme according to the theme.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{select}{An series of string. The first feature of var_select will be chosen from the first experiment, and the second feature will be chosen from the second experiment.}
}
\description{
Two variable regression plot
}
\examples{
data(MAE)
## For EMPT
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_fitline_plot(var_select=c('1.1.1.1','BMI'))

MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_fitline_plot(var_select=c('Weight','BMI'),se=TRUE,
                   estimate_group='Sex',show='html')

## For EMP
k1 <- MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_decostand(method = 'clr')

k2 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(collapse_by='row',estimate_group = 'Class',
               method = 'sum') |>
  EMP_decostand(method = 'relative')

(k1 + k2) |>
  EMP_fitline_plot(var_select=c('1.1.1.1','Bacilli'),
                   estimate_group='Group',eq_size=2.5)
}
