% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{EMP_function_import}
\alias{EMP_function_import}
\title{Title}
\usage{
EMP_function_import(x, type, assay_name = NULL, humann_format = FALSE)
}
\arguments{
\item{x}{wait_for_add}

\item{type}{wait_for_add}

\item{assay_name}{wait_for_add}

\item{humann_format}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
