% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_GSEA.R
\name{EMP_GSEA_analysis}
\alias{EMP_GSEA_analysis}
\title{Gene set enrichment analysis.}
\usage{
EMP_GSEA_analysis(
  x,
  condition,
  experiment,
  estimate_group = NULL,
  method,
  cor_method = "pearson",
  group_level = NULL,
  keyType = NULL,
  KEGG_Type = "KEGG",
  species = "all",
  pseudocount = 1e-04,
  pvalueCutoff = 1,
  threshold = NULL,
  threshold_r = 0,
  threshold_p = 0.05,
  seed = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{condition}{Expressions that return a logical value. The alogarithm condition used in method = "log2FC".eg. pvalue < 0.05}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{estimate_group}{A character string. Select the column you are interested in the coldata.}

\item{method}{A character string. Methods include signal2Noise, cor, log2FC.}

\item{cor_method}{A character string including pearson, spearman, kendall. The alogarithm cor_method used in method = "cor".}

\item{group_level}{A series of character strings. Determine the comparison order of groups when method = "log2FC".}

\item{keyType}{A character string. keyType include ko, ec, cpd, entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{pseudocount}{A number. The alogarithm pseudocount used in method = "signal2Noise", adjust the 0 in the signal2Noise result into pseudocount value. (default:0.0001)}

\item{pvalueCutoff}{A character string. Adjusted pvalue cutoff on enrichment tests to report.}

\item{threshold}{A number. The alogarithm threshold used in method = "signal2Noise",filter out the feature below the signal2Noise threshold.}

\item{threshold_r}{A number. The alogarithm threshold used in method = "cor",filter out the feature below the abusolte corffcient threshold.}

\item{threshold_p}{A number. The alogarithm threshold used in method = "cor",filter out the feature above the cor test pavlue threshold.}

\item{seed}{A boolean.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to clusterProfiler::GSEA.}
}
\value{
EMPT object
}
\description{
Gene set enrichment analysis.
}
\examples{
# add example
}
