% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.EMP.utilis.R
\name{EMP_heatmap_plot,EMP_WGCNA_cor_analysis2-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis2-method}
\title{Title}
\usage{
\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis2}(obj)
}
\arguments{
\item{EMP_WGCNA_cor_analysis2}{}
}
\value{
x object
}
\description{
Title
}
\examples{
# 
}
