% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_heatmap.R
\name{EMP_heatmap_WGCNA}
\alias{EMP_heatmap_WGCNA}
\title{Title}
\usage{
EMP_heatmap_WGCNA(
  x,
  palette = c("steelblue", "white", "darkred"),
  show = "all",
  mytheme = "theme()"
)
}
\arguments{
\item{x}{wait_for_add}

\item{palette}{wait_for_add}

\item{show}{wait_for_add}

\item{mytheme}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
