% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_history}
\alias{EMP_history}
\title{Get history from EMPT or EMP object}
\usage{
EMP_history(obj)
}
\arguments{
\item{obj}{Object in EMPT or EMP format.}
}
\value{
EMPT object
}
\description{
Get history from EMPT or EMP object
}
\examples{
# add example
}
