% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_identify_assay.R
\name{EMP_identify_assay}
\alias{EMP_identify_assay}
\title{Title}
\usage{
EMP_identify_assay(
  x,
  experiment,
  estimate_group = NULL,
  method = c("default", "edgeR"),
  min = if (method == "edgeR") 10 else 0.001,
  min_ratio = 0.7,
  action = "add"
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{method}{wait_for_add}

\item{min}{wait_for_add}

\item{min_ratio}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
