% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_impute.R
\name{EMP_impute}
\alias{EMP_impute}
\title{Title}
\usage{
EMP_impute(
  x,
  experiment,
  coldata = T,
  assay = F,
  rowdata = F,
  pmm.k = 10,
  num.trees = 1000,
  seed = 123,
  verbose = 0,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{coldata}{wait_for_add}

\item{assay}{wait_for_add}

\item{rowdata}{wait_for_add}

\item{pmm.k}{wait_for_add}

\item{num.trees}{wait_for_add}

\item{seed}{wait_for_add}

\item{verbose}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
