% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_marker_analysis.R
\name{EMP_marker_analysis}
\alias{EMP_marker_analysis}
\title{Marker discover based on classify or regression model}
\usage{
EMP_marker_analysis(
  obj,
  experiment,
  method,
  estimate_group,
  seed = 123,
  nfolds = 5,
  lambda_select = "lambda.min",
  max.depth = 6,
  eta = 0.3,
  nrounds = 50,
  objective = "binary:logistic",
  verbose = 0,
  action = "add",
  ...
)
}
\arguments{
\item{obj}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Method includs boruta, randomforest, xgboost, lasso.}

\item{estimate_group}{A character string. Select the group name in the coldata to be calculated.}

\item{seed}{An interger. Set the random seed to the plot.}

\item{nfolds}{An interger. Only actived when method = 'lasso'. More imformation in glmnet::cv.glmnet.}

\item{lambda_select}{A character string including lambda.min or lambda.1se. Only actived when method = 'lasso'. More imformation in glmnet::cv.glmnet.}

\item{max.depth}{An interger (default:6). Only actived when method = 'xgboost'. More imformation in xgboost::xgboost.}

\item{eta}{A number (0.3). Only actived when method = 'xgboost'. More imformation in xgboost::xgboost.}

\item{nrounds}{An interger (default:50). Only actived when method = 'xgboost'. More imformation in xgboost::xgboost.}

\item{objective}{An character string (default:binary:logistic). Only actived when method = 'xgboost'. More imformation in xgboost::xgboost.}

\item{verbose}{An interger (default:0). Only actived when method = 'xgboost'. More imformation in xgboost::xgboost.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function Boruta::Boruta, randomForest, xgboost::xgboost, glmnet::cv.glmnet.}
}
\value{
EMPT object
}
\description{
Marker discover based on classify or regression model
}
\examples{
# add example
}
