% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_message}
\alias{EMP_message}
\title{Info output with style and color}
\usage{
EMP_message(
  str,
  order = 0,
  color = 32,
  pure = FALSE,
  show = "message",
  call. = FALSE,
  ...
)
}
\arguments{
\item{str}{data.}

\item{order}{order.}

\item{color}{color.}

\item{pure}{logical. Default message, warning and stop.}

\item{show}{character including message, warning and stop.}

\item{call.}{logical, indicating if the call should become part of the warning or error message.}

\item{...}{See also \code{\link[base]{message}}, \code{\link[base]{warning}} and \code{\link[base]{stop}}.}
}
\description{
Info output with style and color
}
\section{Detaild }{

This function performed followed by the ANSI.

For color example:

30: Black

31: Red

32: Green

33: Yellow

34: Blue

35: Magenta

36: Cyan

37: White


For order example:

0: Regular

1: Bold

2: blurred

3: Italic

4: Underline

5: Blinking text

9: Strikethrough

40: Black background

41: Red background

42: Green background

43: Yellow background

44: Blue background

45: Magenta background

46: Cyan background

47: White background
}

\examples{
\dontrun{
# blue
EMP_message('Hello World',color = 94,order = 0)
# green 
EMP_message('Hello World',color = 32,order = 0)
# green and bold
EMP_message('Hello World',color = 32,order = 1)
# background  
EMP_message('Hello World',color = 31,order = 7)
# underline 
EMP_message('Hello World',color = 32,order = 4)
# italics 
EMP_message('Hello World',color = 32,order = 3)
# blurred
EMP_message('Hello World',color = 32,order = 2)

# warning with background  
EMP_message('Hello World',color = 31,order = 7,show='warning')

# stop with background  
EMP_message('Hello World',color = 31,order = 7,show='stop')
}
}
\author{
Bingdong Liu
}
