% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_multi.R
\name{EMP_multi_analysis}
\alias{EMP_multi_analysis}
\title{Multi analysis for diff reslut or enrichment}
\usage{
EMP_multi_analysis(
  EMP,
  select = NULL,
  method = "feature",
  combineFun = "enricher",
  pvalueCutoff = 0.05,
  combineMethod = "fisher",
  combineGroup = FALSE,
  combineLevel = "gene",
  p.adjust = "fdr",
  minGSSize = 10,
  maxGSSize = 500,
  keyType,
  KEGG_Type = "KEGG",
  species = "all",
  action = "add",
  ...
)
}
\arguments{
\item{EMP}{EMP object}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string including feature, diff_feature_enrich and same_feature_enrich.}

\item{combineFun}{A character string including enricher, ActivePathways and mitch.}

\item{pvalueCutoff}{A character string. Adjusted pvalue cutoff on enrichment tests to report.}

\item{combineMethod}{A character string including fisher, edgington and stouffer. Only actived, when combineFun == "enricher".}

\item{combineGroup}{A boolean. Whether the function combine the enrichment or not.}

\item{combineLevel}{A character string. One of "gene" and "enrichResult". The parameter options include 'gene' and 'enrichResult' indicate whether the fusion p-value occurs at the feature level or at the enrichment result level. Only actived, when combineFun = "enricher".}

\item{p.adjust}{A character string. One of holm, hochberg, hommel, bonferroni, BH, BY, fdr, none.(defaault: fdr)}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing. (default=10)}

\item{maxGSSize}{Maximal size of each geneSet for analyzing. default=500)}

\item{keyType}{A character string. Methods include ko, ec, cpd and entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{...}
}
\value{
EMPT object
}
\description{
Multi analysis for diff reslut or enrichment
}
\examples{
data(MAE)
\dontrun{
## prepare the result
k1 <- MAE |> 
  EMP_assay_extract('geno_ec') |> 
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group) |> 
  EMP_enrich_analysis(pvalue < 0.05,keyType = 'ec')

## Use different differential analysis methods to mimic the statistical results 
## of the same feature across different cohorts.
k2 <- MAE |>
  EMP_assay_extract('geno_ec') |> 
  EMP_diff_analysis(method = 'wilcox.test',estimate_group = 'Group') |>
  EMP_enrich_analysis(pvalue < 0.05,keyType = 'ec')

k3 <- MAE |> 
  EMP_assay_extract('geno_ko')|> 
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group) |> 
  EMP_enrich_analysis(pvalue < 0.05,keyType = 'ko')

k4 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group) |> 
  EMP_enrich_analysis(pvalue < 0.05,keyType = 'cpd')

## Melt diff analysis of the same feature from two result
(k1+k2) |> 
  EMP_multi_analysis(method = 'feature',combineMethod='edgington',p.adjust = 'BH') 

## Melt diff analysis of the same feature from two result and make a combine enrichment
(k1+k2) |> 
  EMP_multi_analysis(method = 'same_feature_enrich',keyType = 'ec',combineFun='ActivePathways')  

## Melt diff analysis of the different feature from multi-results and make a combine enrichment
(k1+k4) |> 
  EMP_multi_analysis(method = 'diff_feature_enrich')  

## Visualization
(k1+k2) |>
  EMP_multi_analysis(method = 'same_feature_enrich',keyType = 'ec',combineFun='ActivePathways') |>
  EMP_enrich_dotplot()
}
}
