% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_mutate.R
\name{EMP_mutate}
\alias{EMP_mutate}
\title{Create, modify, and delete columns for EMPT object}
\usage{
EMP_mutate(
  obj,
  experiment = NULL,
  ...,
  .by = NULL,
  .before = NULL,
  .after = NULL,
  mutate_by = "sample",
  location = "coldata",
  action = "colwise",
  keep_result = FALSE,
  use_cached = TRUE
)
}
\arguments{
\item{obj}{MAE or EMPT object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{...}{... Addtional parameters, see also \code{\link[dplyr]{mutate}}.}

\item{.by}{Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by(). See also \code{\link[dplyr]{mutate}}.}

\item{.before}{Optionally, control where new columns should appear.}

\item{.after}{(the default is to add to the right hand side). See relocate() for more details.}

\item{mutate_by}{A character string inluding sample and feature.}

\item{location}{A character string inluding assay, rowdata and coldata.}

\item{action}{A character string inluding colwise and rowwise, indicating whether operations should be performed column-wise or row-wise.}

\item{keep_result}{If the input is TRUE, it means to keep all analysis results,regardless of how samples and features change. If the input is a name, it means to keep the corresponding analysis results.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}
}
\value{
EMPT object
}
\description{
Create, modify, and delete columns for EMPT object
}
\section{Detalis}{

Addtional parameters inherits from the function mutate from package "dplyr", but the parameters .keep is not allowed to reset.
}

\examples{
# Create, modify, and delete the features in the assay according the data and results
data(MAE)

# Create, modify, and delete the features in the assay according the data and results
# Caculte the F/B ratio in the assay
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(collapse_by = 'row',estimate_group = 'Phylum') |>
  EMP_mutate(F_B_ratio = Firmicutes/Bacteroidetes,
             .before = 2,
             mutate_by = 'sample',location ='assay') |>
  EMP_filter(filterFeature = 'F_B_ratio') |>
  EMP_boxplot(estimate_group='Group')

# Add the new value in the rowdata
# Combine the different pvalue
MAE |>
  EMP_assay_extract('host_gene') |>
  EMP_diff_analysis(method = 'wilcox.test',estimate_group = 'Group') |>
  EMP_mutate(wilcox = pvalue,
             wilcox_fdr = fdr,
             mutate_by = 'feature',location ='rowdata') |>
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group) |>
  EMP_mutate(DESeq2 = pvalue,
             DESeq2_fdr = fdr,
             mutate_by = 'feature',location ='rowdata')

# Add the new value in the coldata
# Create the new group
MAE |>
  EMP_assay_extract('host_gene',pattern = 'A1') |>
  EMP_mutate(Degree = dplyr::case_when(
    BMI < 18.5                        ~ "Lean",
    BMI >= 18.5 & BMI < 24            ~ "Normal",
    BMI >= 24 & BMI < 28              ~ "Fat",
    TRUE                              ~ "Need Med"
    ),
    mutate_by = 'sample',location = 'coldata',.after = Group
  ) |>
  EMP_boxplot(estimate = 'Degree')

# Change the whole coldata for the MultiAssayExperiment
MAE |>
  EMP_mutate(Degree = dplyr::case_when(
    BMI < 18.5                        ~ "Lean",
    BMI >= 18.5 & BMI < 24            ~ "Normal",
    BMI >= 24 & BMI < 28              ~ "Fat",
    TRUE                              ~ "Need Med"),
    .after = Group) |>
  EMP_coldata_extract()
}
