% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_tidy_method.R
\name{EMP_mutate_deprecated}
\alias{EMP_mutate_deprecated}
\title{Title}
\usage{
EMP_mutate_deprecated(x, experiment = NULL, ...)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
