% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{EMP_netplot}
\alias{EMP_netplot}
\alias{EMP_netplot,EMP_enrich_analysis-method}
\title{Title}
\usage{
EMP_netplot(obj, ...)

\S4method{EMP_netplot}{EMP_enrich_analysis}(obj, ...)
}
\arguments{
\item{obj}{obj}

\item{...}{...}

\item{EMP_enrich_analysis}{xx}
}
\value{
xx object

xx object
}
\description{
Title

Title
}
\examples{
# 
#
}
