% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_EMP_network_plot.R
\name{EMP_network_plot}
\alias{EMP_network_plot}
\title{Network plot}
\usage{
EMP_network_plot(
  obj,
  layout = "spring",
  weighted = TRUE,
  threshold = 0,
  shape = "circle",
  node_color = c("#F0E442", "#009E73", "#E69F00", "#CC79A7", "#8491B4FF", "#F39B7FFF",
    "#4DBBD5FF", "#E64B35FF", "#D55E00", "#00A087FF", "#56B4E9", "#B2182B"),
  node_info = NULL,
  vsize = NULL,
  esize = NULL,
  posCol = "#B2182B",
  negCol = "#002FA7",
  fade = FALSE,
  negDashed = TRUE,
  label.cex = 0.4,
  label.color = "black",
  label.prop = 0.9,
  edge.labels = FALSE,
  edge.label.cex = 0.5,
  legend.cex = 0.3,
  legend.mode = NULL,
  show = "net",
  ...
)
}
\arguments{
\item{obj}{EMPT or EMP object.}

\item{layout}{Placement of the nodes including spring and circle.}

\item{weighted}{A  Logical, should the analyzed network be weighted?}

\item{threshold}{A numeric value that defaults to 0. Edges with absolute weight that are not above this value are REMOVED from the network.}

\item{shape}{A character containing the shape of the nodes. "circle", "square", "triangle" and "diamond" are supported.}

\item{node_color}{A vector with a color for each element in the groups list.}

\item{node_info}{A character in the rowdata to display more info for node.}

\item{vsize}{A value indicating the size of the nodes.}

\item{esize}{Size of the largest edge.}

\item{posCol}{Color of positive edges.}

\item{negCol}{Color of negative edges.}

\item{fade}{if TRUE (default) and if 'edge.color' is assigned, transparency will be added to edges that are not transparent (or for which no transparency has been assigned) relative to the edge strength, similar if 'trans' is set to TRUE.}

\item{negDashed}{Logical, set to TRUE to make negative edges dashed (overwrites lty).}

\item{label.cex}{Scalar on the label size.}

\item{label.color}{Character containing the color of the labels, defaults to "black"}

\item{label.prop}{Controls the proportion of the width of the node that the label rescales to. Defaults to 0.9.}

\item{edge.labels}{If FALSE, no edge labels are plotted. If TRUE, numerical edge weights are printed on the edges.}

\item{edge.label.cex}{Either a single number or a number per edge used as a scalar of the edge label size. Defaults to 0.5.}

\item{legend.cex}{Scalar of the legend. defaults to 0.4.}

\item{legend.mode}{Character string indicating the type of legend to be drawn. "groups" indicates the legend should be based on the groups object, "names" indicates the legend should be based on the nodeNames object, and style1 and style2 indicate the legend should be based on both. Defaults to "style1" if both "groups" and "nodeNames" arguments are used.}

\item{show}{A character string include net (default), node.}

\item{...}{Additional parameters, see also \code{\link[bootnet]{plot.bootnetResult}} and \code{\link[qgraph]{qgraph}}.}
}
\description{
Network plot
}
\examples{
data(MAE)

# For single experiment
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='wilcox.test', estimate_group = 'Group') |>
  EMP_filter(feature_condition = pvalue<0.05) |>
  EMP_network_analysis(coldata_to_assay = c('BMI','PHQ9','GAD7')) |>
  EMP_network_plot(node_info = 'Class',label.cex = 1,edge.labels = TRUE)

# For muti experiment
k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='wilcox.test', estimate_group = 'Group') |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)

(k1 + k2 ) |> 
  EMP_network_analysis() |> 
  EMP_network_plot(show = 'net',layout = 'spring',
                   edge.labels=TRUE,edge.label.cex=0.5,
                   vsize = 5,threshold = 0.4,
                   node_info = c('Phylum','MS2class'),
                   legend.cex=0.3,label.cex = 1,label.prop = 0.9,
                   fade = TRUE,shape = 'circle',font=2)
(k1 + k2 ) |> 
  EMP_network_analysis() |> 
  EMP_network_plot(show = 'node') # get the node importance
}
