% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{EMP_normal_import}
\alias{EMP_normal_import}
\title{Title}
\usage{
EMP_normal_import(
  x,
  sampleID = NULL,
  dfmap = NULL,
  assay_name = NULL,
  assay = NULL
)
}
\arguments{
\item{x}{wait_for_add}

\item{sampleID}{wait_for_add}

\item{dfmap}{wait_for_add}

\item{assay_name}{wait_for_add}

\item{assay}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# example
}
