% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_pca.R
\name{EMP_pca}
\alias{EMP_pca}
\title{Title}
\usage{
EMP_pca(
  data,
  seed = 123,
  group_level = c("default"),
  method = c("t.test"),
  key_samples = NULL,
  ellipse = NULL,
  width = 15,
  height = 15,
  distance = c("bray"),
  palette = c("#E64B35FF", "#4DBBD5FF", "#00A087FF", "#3C5488FF", "#F39B7FFF",
    "#8491B4FF", "#B2182B", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7", "#CC6666")
)
}
\arguments{
\item{data}{wait_for_add}

\item{seed}{wait_for_add}

\item{group_level}{wait_for_add}

\item{method}{wait_for_add}

\item{key_samples}{wait_for_add}

\item{ellipse}{wait_for_add}

\item{width}{wait_for_add}

\item{height}{wait_for_add}

\item{distance}{wait_for_add}

\item{palette}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
