% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_pca.R
\name{EMP_pca_plot}
\alias{EMP_pca_plot}
\title{Title}
\usage{
EMP_pca_plot(pca_obj, plot)
}
\arguments{
\item{pca_obj}{wait_for_add}

\item{plot}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
