% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMP_prepare_data}
\alias{EMP_prepare_data}
\title{EMP_prepare_data}
\format{
list
}
\usage{
EMP_prepare_data
}
\value{
list
}
\description{
This data set contains the EMP_prepare_data data.
}
\keyword{datasets}
