% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{EMP_RFCV}
\alias{EMP_RFCV}
\title{Title}
\usage{
EMP_RFCV(
  RF,
  seed_start = 123,
  ntree = 1000,
  kfold = 5,
  rep = 10,
  RF_importance = 1,
  step = 1,
  each_ouput = F,
  x_break = 1,
  cutoff_colour = c("red"),
  palette = c("#E64B35FF", "#4DBBD5FF", "#00A087FF", "#3C5488FF", "#F39B7FFF",
    "#8491B4FF", "#B2182B", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7", "#CC6666")
)
}
\arguments{
\item{RF}{wait_for_add}

\item{seed_start}{wait_for_add}

\item{ntree}{wait_for_add}

\item{kfold}{wait_for_add}

\item{rep}{wait_for_add}

\item{RF_importance}{wait_for_add}

\item{step}{wait_for_add}

\item{each_ouput}{wait_for_add}

\item{x_break}{wait_for_add}

\item{cutoff_colour}{wait_for_add}

\item{palette}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
