% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_save_var}
\alias{EMP_save_var}
\title{Save result into GlobalEnv}
\usage{
EMP_save_var(data, var_name, ...)
}
\arguments{
\item{data}{data}

\item{var_name}{the name of data in the GlobalEnv}

\item{...}{Further parameters passed to assign}
}
\value{
EMPT object
}
\description{
Save result into GlobalEnv
}
\examples{
data(MAE)
MAE |> 
  EMP_assay_extract('host_gene',pattern = 'A1BG',pattern_ref = 'feature') |>
  EMP_save_var('temp_data') |>
  EMP_collapse(estimate_group = 'Group',collapse_by = 'col') |>
  EMP_heatmap_plot()

temp_data
}
