% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_scatterplot_reduce_dimension.R
\name{EMP_scatterplot_reduce_dimension}
\alias{EMP_scatterplot_reduce_dimension}
\title{Title}
\usage{
EMP_scatterplot_reduce_dimension(
  EMPT,
  seed = 123,
  group_level = "default",
  show = "p12",
  distance_for_adonis = NULL,
  force_adonis = FALSE,
  estimate_group = NULL,
  palette = NULL,
  method = "t.test",
  key_samples = NULL,
  ellipse = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{seed}{wait_for_add}

\item{group_level}{wait_for_add}

\item{show}{wait_for_add}

\item{distance_for_adonis}{wait_for_add}

\item{force_adonis}{force_adonis}

\item{estimate_group}{wait_for_add}

\item{palette}{wait_for_add}

\item{method}{wait_for_add}

\item{key_samples}{wait_for_add}

\item{ellipse}{wait_for_add}

\item{width}{wait_for_add}

\item{height}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
