% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_EMP_structure_plot.R
\name{EMP_structure_plot}
\alias{EMP_structure_plot}
\title{Structure plot for EMPT result}
\usage{
EMP_structure_plot(
  obj,
  plot_category = 1,
  method = "mean",
  top_num = 10,
  estimate_group = NULL,
  bar_level = "default",
  ncol = NULL,
  show = "pic",
  palette = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{obj}{EMPT object}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{method}{A character string including mean,median,max and min. The name of the statistical test that is applied to select the top feature.}

\item{top_num}{An interger.}

\item{estimate_group}{A character string. Select the colname in the coldata to divide the data in the plot.}

\item{bar_level}{A series of character string. Reset the bar order.}

\item{ncol}{An interger. Set the col number in the facet plot.}

\item{show}{A character string include pic (default), html(under constrution).}

\item{palette}{A series of character string. Color palette.}

\item{mytheme}{Modify components of a theme according to the \code{\link[ggplot2]{theme}} and \code{\link[ggplot2]{ggtheme}}.}
}
\value{
EMPT object
}
\description{
Structure plot for EMPT result
}
\examples{
data(MAE)
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_decostand(method = 'relative') |>
  EMP_collapse(estimate_group = 'Class',collapse_by = 'row') |>
  EMP_structure_plot(top_num=10)
# merge the data by group information 
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_decostand(method = 'relative') |>
  EMP_collapse(estimate_group = 'Class',collapse_by = 'row') |>
  EMP_collapse(estimate_group = 'Group',collapse_by = 'col') |>
  EMP_structure_plot(top_num=10)

# plot by group information
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_decostand(method = 'relative') |>
  EMP_collapse(estimate_group = 'Class',collapse_by = 'row') |>
  EMP_structure_plot(top_num=10,estimate_group='Sex')
    
MAE |>
  EMP_assay_extract('host_gene') |>
  EMP_identify_assay(method = 'edgeR') |>
  EMP_structure_plot(top_num=10)
}
