% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_summary.R
\name{EMP_summary}
\alias{EMP_summary}
\title{Visual charts for MultiAssayExperiment or EMP object}
\usage{
EMP_summary(obj)
}
\arguments{
\item{obj}{MultiAssayExperiment or EMP object}
}
\value{
view object
}
\description{
Visual charts for MultiAssayExperiment or EMP object
}
\examples{
# add example
}
