% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_import.R
\name{EMP_taxonomy_import}
\alias{EMP_taxonomy_import}
\title{Import microbial data into SummariseExperiment}
\usage{
EMP_taxonomy_import(
  file = NULL,
  data = NULL,
  humann_format = FALSE,
  assay_name = NULL,
  duplicate_feature = FALSE,
  sep = if (humann_format == "FALSE") ";" else "|"
)
}
\arguments{
\item{file}{A file path.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{humann_format}{A boolean. Whether the function improt the data according to the humann format.}

\item{assay_name}{A character string. Indicate what kind of result the data belongs to, such as counts, relative abundance, TPM, etc.}

\item{duplicate_feature}{A boolean. Whether the feature exist the dupicated name.}

\item{sep}{The field separator character. Values on feature column of the file are separated by this character. (defacult:';',when humann_format=T defacult:'|')}
}
\value{
SummmariseExperiment object
}
\description{
Import microbial data into SummariseExperiment
}
\examples{
# add example
}
