% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_to_EMP1}
\alias{EMP_to_EMP1}
\title{Transfer microbial data from EasyMultiProfiler to EasyMciroPlot}
\usage{
EMP_to_EMP1(obj, estimate_group)
}
\arguments{
\item{obj}{EMPT object.}

\item{estimate_group}{A character string. Select the column in the coldata to add in the mapping file.}
}
\value{
list
}
\description{
Transfer microbial data from EasyMultiProfiler to EasyMciroPlot
}
\examples{
data(MAE)
# Get the data from EasyMultiProfiler
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_feature_convert(from = 'tax_single',add = 'tax_full') |>
  EMP_to_EMP1(estimate_group = 'Group') -> deposit
\dontrun{ 
# Work in the EasyMicroPlot
library(EasyMicroPlot)
alpha_re <- alpha_plot(data = deposit$data,
                       design = deposit$mapping,
                       min_relative = 0.001,min_ratio = 0.7,method = 'ttest')
}
}
