% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_WGCNA_analysis.R
\name{EMP_WGCNA_cluster_analysis}
\alias{EMP_WGCNA_cluster_analysis}
\title{Title}
\usage{
EMP_WGCNA_cluster_analysis(
  x,
  experiment,
  use_cached = T,
  powers = c(1:10, seq(from = 12, to = 20, by = 2)),
  RsquaredCut = 0.85,
  removeFirst = FALSE,
  nBreaks = 10,
  blockSize = NULL,
  corFnc = WGCNA::cor,
  corOptions = list(use = "p"),
  networkType = "unsigned",
  moreNetworkConcepts = FALSE,
  gcInterval = NULL,
  TOMType = "unsigned",
  minModuleSize = 30,
  reassignThreshold = 0,
  mergeCutHeight = 0.25,
  numericLabels = TRUE,
  pamRespectsDendro = FALSE,
  saveTOMs = T,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{powers}{wait_for_add}

\item{RsquaredCut}{wait_for_add}

\item{removeFirst}{wait_for_add}

\item{nBreaks}{wait_for_add}

\item{blockSize}{wait_for_add}

\item{corFnc}{wait_for_add}

\item{corOptions}{wait_for_add}

\item{networkType}{wait_for_add}

\item{moreNetworkConcepts}{wait_for_add}

\item{gcInterval}{wait_for_add}

\item{TOMType}{wait_for_add}

\item{minModuleSize}{wait_for_add}

\item{reassignThreshold}{wait_for_add}

\item{mergeCutHeight}{wait_for_add}

\item{numericLabels}{wait_for_add}

\item{pamRespectsDendro}{wait_for_add}

\item{saveTOMs}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
