% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.EMP.utilis.R
\name{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis,EMPT-method}
\alias{EMP_WGCNA_cor_analysis,EMP-method}
\title{EMP_WGCNA_cor_analysis}
\usage{
EMP_WGCNA_cor_analysis(obj, ...)

\S4method{EMP_WGCNA_cor_analysis}{EMPT}(
  obj,
  method = "spearman",
  coldata_to_assay = NULL,
  action = "add",
  ...
)

\S4method{EMP_WGCNA_cor_analysis}{EMP}(
  obj,
  select = NULL,
  method = "spearman",
  action = "add",
  ...
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{...}{Further parameters passed to the function agricolae::correlation}

\item{method}{A character string. Methods include pearson (default), spearman and kendall.}

\item{coldata_to_assay}{A series of character strings. Select the column from coldata to caculate.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{select}{A character string. The experiment name in the EMP object.}
}
\value{
EMP object
}
\description{
EMP_WGCNA_cor_analysis
}
\examples{
#x
}
