% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{EMPT-class}
\alias{EMPT-class}
\title{Class "EMPT"
This class represents the ...}
\description{
Class "EMPT"
This class represents the ...
}
\section{Slots}{

\describe{
\item{\code{deposit}}{deposit}

\item{\code{deposit2}}{deposit2}

\item{\code{plot_deposit}}{plot_deposit}

\item{\code{deposit_append}}{deposit_append}

\item{\code{deposit_info}}{deposit_info}

\item{\code{experiment}}{experiment}

\item{\code{assay_name}}{assay_name}

\item{\code{estimate_group}}{estimate_group}

\item{\code{estimate_group_info}}{estimate_group_info}

\item{\code{message_info}}{message_info}

\item{\code{formula}}{formula}

\item{\code{method}}{method}

\item{\code{algorithm}}{algorithm}

\item{\code{history}}{history}

\item{\code{palette}}{palette}

\item{\code{plot_category}}{plot_category}

\item{\code{plot_specific}}{plot_specific}

\item{\code{plot_info}}{plot_info}

\item{\code{info}}{info}
}}

\author{
Bingdong Liu
}
\keyword{classes}
