% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_convert.R
\name{feature_convert}
\alias{feature_convert}
\title{Title}
\usage{
feature_convert(
  feature,
  from = "SYMBOL",
  to = "ENTREZID",
  species = "Human",
  OrgDb = NULL
)
}
\arguments{
\item{feature}{features}

\item{from}{from}

\item{to}{to}

\item{species}{speckes}

\item{OrgDb}{OrgDb}
}
\value{
data.frame
}
\description{
Title
}
\examples{
#
}
