% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{get_Xd}
\alias{get_Xd}
\title{get_Xd}
\usage{
get_Xd(distance_score, n, pia, genelist, TERM2GENE)
}
\arguments{
\item{distance_score}{distance_score}

\item{n}{n}

\item{pia}{pia}

\item{genelist}{genelist}

\item{TERM2GENE}{TERM2GENE}
}
\description{
get_Xd
}
