% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{get_Xd2}
\alias{get_Xd2}
\title{get_Xd2}
\usage{
get_Xd2(distance_score, n, genelist, u)
}
\arguments{
\item{distance_score}{distance_score}

\item{n}{n}

\item{genelist}{genelist}

\item{u}{u}
}
\description{
get_Xd2
}
