% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_gason.R
\name{gson_cpd2}
\alias{gson_cpd2}
\title{Title}
\usage{
gson_cpd2(KEGG_Type = "KEGG", species = "all")
}
\arguments{
\item{KEGG_Type}{KEGG_Type}

\item{species}{species}
}
\value{
gson object
}
\description{
Title
}
\examples{
# xx
}
