% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_gason.R
\name{gson_KEGG2}
\alias{gson_KEGG2}
\title{Title}
\usage{
gson_KEGG2(keyType = "ko", KEGG_Type = "KEGG", species = "all")
}
\arguments{
\item{keyType}{keyType}

\item{KEGG_Type}{keyType}

\item{species}{keyType}
}
\value{
gson object
}
\description{
Title
}
\examples{
# xx
}
