% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{humann_function_import}
\alias{humann_function_import}
\title{Title}
\usage{
humann_function_import(x, type)
}
\arguments{
\item{x}{wait_for_add}

\item{type}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
