% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_import.R
\name{humann_taxonomy_import}
\alias{humann_taxonomy_import}
\title{Title}
\usage{
humann_taxonomy_import(file = NULL, data = NULL, sep = "|")
}
\arguments{
\item{file}{A file path.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{sep}{The field separator character. Values on each line of the file are separated by this character. (defacult:'|')}
}
\value{
SummmariseExperiment object
}
\description{
Title
}
\examples{
# add example
}
