% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_list2dataframe.R
\name{list2dataframe}
\alias{list2dataframe}
\title{Transform the result data of gene differential expression from list to dataframe}
\usage{
list2dataframe(genelists, method = "intersection")
}
\arguments{
\item{genelists}{Results of gene difference analysis.}

\item{method}{The method of merging, one of "intersection" and "union"}
}
\description{
Transform the result data of gene differential expression from list to dataframe
}
\examples{
\dontrun{
    library(multiGSEA)
    data(transcriptome)
    data(proteome)
    data(metabolome) 
    class(transcriptome) <- class(proteome) <- class(metabolome) <- "data.frame"
    transcriptome2 <- transcriptome$pValue
    names(transcriptome2) <- transcriptome$Symbol
    proteome2 <- proteome$pValue
    names(proteome2) <- proteome$Symbol
    proteome2 <- proteome2[-which(names(proteome2)== "")]
    genelists <- list(transcriptome2 = transcriptome2, proteome2 = proteome2)
    geneDf <- list2dataframe(genelists)
}
}
