% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metaboliteIDmapping_data}
\alias{metaboliteIDmapping_data}
\title{a data.frame of xx}
\format{
A data.frame with xx
}
\usage{
metaboliteIDmapping_data
}
\value{
A dataframe
}
\description{
xx
}
\keyword{datasets}
