% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment.R
\name{multiEnrichment}
\alias{multiEnrichment}
\title{Multi-omics enrichment analysis}
\usage{
multiEnrichment(
  multiGene,
  method = "enricher",
  network = NULL,
  cutoff = 0.05,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  TERM2GENE,
  TERM2NAME = NULL,
  combineMethod = "fisher",
  stoufferWeights = NULL,
  combineLevel = "enrichResult",
  ...
)
}
\arguments{
\item{multiGene}{a data.frame of multi-omics gene difference analysis results (pvalue).
Each row is a gene, and each column represents an omics dataset.}

\item{method}{enrichment analysis method, one of "enricher"(the default)
"GSEA", "mitch", "ActivePathways", and "multiNetEnrich".}

\item{network}{network}

\item{cutoff}{Pvalue threshold of differentially expressed genes.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", 
"bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{qvalueCutoff}{Cutoff of qvalue.}

\item{TERM2GENE}{user input annotation of TERM TO GENE mapping, 
a data.frame of 2 column with term and gene}

\item{TERM2NAME}{user input of TERM TO NAME mapping, 
a data.frame of 2 column with term and name}

\item{combineMethod}{The method of combining pvalues, one of 
"fisher", "edgington", "stouffer" and "Brown"(only used in ActivePathways method).}

\item{stoufferWeights}{weights of stouffer combine method.}

\item{combineLevel}{one of "gene" and "enrichResult"}

\item{...}{Other parameters.}
}
\description{
Multi-omics enrichment analysis
}
