% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_class2clusterProfiler.R
\name{multiEnrichResult2enrichResult}
\alias{multiEnrichResult2enrichResult}
\title{multiEnrichResult2enrichResult}
\usage{
multiEnrichResult2enrichResult(em_enricher, method = "union")
}
\arguments{
\item{em_enricher}{em_enricher}

\item{method}{method}
}
\description{
multiEnrichResult2enrichResult
}
