% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{NetEnrich}
\alias{NetEnrich}
\title{NetEnrich}
\usage{
NetEnrich(
  genelist,
  network,
  p = 0,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  threshold = 1e-09,
  n = 10,
  nperm = 100,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2
)
}
\arguments{
\item{genelist}{a vector of gene weights, which names are genes.}

\item{network}{network}

\item{p}{restart probability}

\item{TERM2GENE}{TERM2GENE}

\item{TERM2NAME}{TERM2NAME}

\item{threshold}{threshold}

\item{n}{n}

\item{nperm}{Number of permutations to do.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", 
"bonferroni", "BH", "BY", "fdr", "none"}

\item{qvalueCutoff}{Cutoff of qvalue.}
}
\description{
NetEnrich
}
