% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{NetEnrich_rank}
\alias{NetEnrich_rank}
\title{NetEnrich_rank}
\usage{
NetEnrich_rank(
  genelist,
  network,
  p = 0,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  threshold = 1e-09,
  n = 10,
  nperm = 100
)
}
\arguments{
\item{genelist}{genelist}

\item{network}{network}

\item{p}{restart probability}

\item{TERM2GENE}{TERM2GENE}

\item{TERM2NAME}{TERM2NAME}

\item{threshold}{threshold}

\item{n}{n}

\item{nperm}{Number of permutations to do.}
}
\description{
Imitate ActivePathways
}
