% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_NetPEA.R
\name{NetPEA2}
\alias{NetPEA2}
\title{NetPEA2 weighted NetPEA(EnrichNet2)}
\usage{
NetPEA2(
  genelist,
  network,
  p = 0.5,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  threshold = 1e-09,
  nperm = 100,
  keepAll = FALSE,
  nperm_method = 1
)
}
\arguments{
\item{genelist}{genelist.}

\item{network}{network.}

\item{p}{restart probability.}

\item{TERM2GENE}{TERM2GENE.}

\item{TERM2NAME}{TERM2NAME.}

\item{threshold}{threshold.}

\item{nperm}{Number of permutations to do.}

\item{keepAll}{If FALSE, just keep pathways that have overlap with input genes.}

\item{nperm_method}{If 1, random genelist, if 2, random genelist and network.}
}
\description{
NetPEA2 weighted NetPEA(EnrichNet2)
}
