% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{+.EMP}
\alias{+.EMP}
\title{Title}
\usage{
\method{+}{EMP}(e1, e2, ...)
}
\arguments{
\item{e1}{An object of class EMP.}

\item{e2}{An object of class EMPT or EMP.}

\item{...}{...}
}
\description{
Title
}
