% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{+.EMPT}
\alias{+.EMPT}
\title{Title}
\usage{
\method{+}{EMPT}(e1, e2, ...)
}
\arguments{
\item{e1}{An object of class EMPT.}

\item{e2}{An object of class EMPT or EMP.}

\item{...}{...}
}
\description{
Title
}
