% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{progress_text}
\alias{progress_text}
\title{Title}
\usage{
progress_text(style = 3, ...)
}
\arguments{
\item{style}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
