% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{RFCV_data_binary}
\alias{RFCV_data_binary}
\title{Title}
\usage{
RFCV_data_binary(data, rf_estimate_group, id_not = "NOT")
}
\arguments{
\item{data}{wait_for_add}

\item{rf_estimate_group}{wait_for_add}

\item{id_not}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
