% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{RFCV_plot}
\alias{RFCV_plot}
\title{Title}
\usage{
RFCV_plot(
  data,
  x_break = 1,
  cutoff_colour = c("red"),
  palette = c("#E64B35FF", "#4DBBD5FF", "#00A087FF", "#3C5488FF", "#F39B7FFF",
    "#8491B4FF", "#B2182B", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7", "#CC6666")
)
}
\arguments{
\item{data}{wait_for_add}

\item{x_break}{wait_for_add}

\item{cutoff_colour}{wait_for_add}

\item{palette}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
