% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_RFCV.R
\name{RFCV_roc}
\alias{RFCV_roc}
\title{Title}
\usage{
RFCV_roc(x, rf_estimate_group, rf_tax_select, ntree = 1000)
}
\arguments{
\item{x}{wait_for_add}

\item{rf_estimate_group}{wait_for_add}

\item{rf_tax_select}{wait_for_add}

\item{ntree}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
