% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{RWR}
\alias{RWR}
\title{Random Walk with Restart}
\usage{
RWR(genelist, network, p = 0.5, threshold = 1e-09)
}
\arguments{
\item{genelist}{a vector of -log(pvalue), with the names of gene id.}

\item{network}{network}

\item{p}{restart probability}

\item{threshold}{threshold}
}
\description{
Random Walk with Restart
}
