% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_show.R
\name{show,EMPT-method}
\alias{show,EMPT-method}
\title{Title}
\usage{
\S4method{show}{EMPT}(object)
}
\arguments{
\item{object}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
