% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{stat_test}
\alias{stat_test}
\title{Perform Differential analysis for table}
\usage{
stat_test(
  data,
  estimate_group = NULL,
  compare_group = NULL,
  formula = NULL,
  value = NULL,
  method = "wilcox.test",
  paired = FALSE,
  paired_group = NULL,
  p.adjust = "fdr",
  ref.group = NULL,
  comparisons = NULL,
  facet.by = NULL,
  error_handling = "warn",
  silent = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame.}

\item{estimate_group}{estimate_group.The primary grouping variable for estimation and comparison.Serves as the comparison group by default.}

\item{compare_group}{compare_group.Optional subgroup analysis variable. When specified, the subgroup comparison performs.}

\item{formula}{formula.}

\item{value}{value}

\item{method}{statistics including t.test, wilcox.test, sign.test, emmeans.test, dunn.test, tukey.hsd.}

\item{paired}{a logical indicating whether you want a paired t-test for t.test and wilcox.test.}

\item{paired_group}{A character string. Variable name corresponding to paired primary or sample.}

\item{p.adjust}{a character string. Adjust P-values for Multiple Comparisons inluding fdr, holm, hochberg, hommel, bonferroni, BH, BY. (default:fdr)}

\item{ref.group}{a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).}

\item{comparisons}{a list of length-2 vectors, which only support t.test, wilcox.test, sign.test and emmeans.test.}

\item{facet.by}{a column name for multiple features.}

\item{error_handling}{error_handling including stop warn and silent.}

\item{silent}{silent.}

\item{...}{additional parameters, see also \code{\link[rstatix]{t_test}},\code{\link[rstatix]{wilcox_test}},...}
}
\description{
Perform Differential analysis for table
}
\examples{
data(ToothGrowth)

# for group
stat_test(
  data = ToothGrowth,
  estimate_group = 'dose',
  value = 'len',
  method = "t.test"
)


stat_test(
  data = ToothGrowth,
  formula = len ~ dose,
  method = "t.test"
)

# for subgroup
stat_test(
  data = ToothGrowth,
  value = 'len',
  estimate_group = "supp",
  compare_group = "dose",
  method = "t.test"
)

stat_test(
  data = ToothGrowth,
  estimate_group = 'supp',
  formula = len ~ dose,
  method = "t.test"
)

 # for ggplot2 
 stat_result <- stat_test(
   data = ToothGrowth,
   formula = len ~ dose,
   value = 'len',
   method = "tukey.hsd",
) 
 
 # for ggpubr
 if(require("ggpubr")){
  ggboxplot(ToothGrowth, x = "dose", y = "len") +
    stat_pvalue_manual(stat_result, label = "p", tip.length = 0.01)
 }

}
