% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{top_detect}
\alias{top_detect}
\title{Detect the top or bottom number}
\usage{
top_detect(x, n, type = "top", index = TRUE)
}
\arguments{
\item{x}{vec.}

\item{n}{a number above 0.}

\item{type}{top or bottom.}

\item{index}{output the index or the number.}
}
\description{
Detect the top or bottom number
}
\examples{
top_detect(1:10,n = 3,index = FALSE)

top_detect(1:10,n = 3,index = TRUE)

MAE |>
  EMP_assay_extract(experiment='taxonomy')|> 
  EMP_alpha_analysis() |>
  EMP_filter(top_detect(shannon,3))
}
\author{
Bingdong Liu
}
