% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{wide2long}
\alias{wide2long}
\title{wide2long}
\usage{
wide2long(network, direct = FALSE)
}
\arguments{
\item{network}{network}

\item{direct}{if TRUE, build a directed network}
}
\description{
wide2long
}
