% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_EnrichNet.R
\name{wide2long_spMatrix}
\alias{wide2long_spMatrix}
\title{wide2long undirected spMatrix network}
\usage{
wide2long_spMatrix(network)
}
\arguments{
\item{network}{network}
}
\description{
wide2long undirected spMatrix network
}
