% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_gason.R
\name{build_gson}
\alias{build_gson}
\title{Build the KEGG gason file including KO, EC, compound.}
\usage{
build_gson(keyType = "ko", KEGG_Type = "KEGG", species = "hsa")
}
\arguments{
\item{keyType}{A character string. keyType include ko, ec, cpd, entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}
}
\value{
gson object
}
\description{
Build the KEGG gason file including KO, EC, compound.
}
\examples{
# xx
}
