% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{.pattern_Dectect_multi}
\alias{.pattern_Dectect_multi}
\title{Title}
\usage{
.pattern_Dectect_multi(pattern_ref, pattern, exact = F)
}
\arguments{
\item{pattern_ref}{Object in EMPT format.}

\item{pattern}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{exact}{A boolean. Whether the extract use exact search method.}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
