% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_adjust_abudance.R
\name{EMP_adjust_abudance}
\alias{EMP_adjust_abudance}
\title{Title}
\usage{
EMP_adjust_abudance(
  x,
  experiment,
  method = "combat_seq",
  use_cached = T,
  .factor_unwanted,
  .factor_of_interest,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or SummarisedExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Methods include combat_seq (default), combat and limma_remove_batch_effect.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{.factor_unwanted}{A tidy select, e.g. column names without double quotation. c(batch, country) These are the factor that we want to adjust for, including unwanted batcheffect, and unwanted biological effects.}

\item{.factor_of_interest}{A tidy select, e.g. column names without double quotation. c(treatment) These are the factor that we want to preserve.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function tidybulk::adjust_abundance.}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
