% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyisis_EMP_alpha.R
\name{EMP_alpha_analysis}
\alias{EMP_alpha_analysis}
\title{Diversity Indices}
\usage{
EMP_alpha_analysis(x, experiment, use_cached = TRUE, action = "add", ...)
}
\arguments{
\item{x}{Object in EMPT format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function vegan::diversity}
}
\value{
EMPT object
}
\description{
Diversity Indices
}
\examples{
# add example
}
