% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_EMP_extract.R
\name{EMP_assay_extract}
\alias{EMP_assay_extract}
\title{Extract assay data}
\usage{
EMP_assay_extract(
  obj,
  experiment,
  pattern_ref = "Name",
  pattern = NULL,
  use_cached = TRUE,
  exact = FALSE,
  action = "add"
)
}
\arguments{
\item{obj}{MultiAssayExperiment or EMPT object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{pattern_ref}{A character string. Select which column in the rowdata to extract assay data from.}

\item{pattern}{A character string. Select which pattern in the rowdata to extract assay data.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{exact}{A boolean. Whether the extract use exact search method.(default:FALSE)}

\item{action}{A character string. A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Extract assay data
}
\examples{
data(MAE)
## Extract the assay data of one existed experiment from MultiAssaayExperiment
MAE |>
 EMP_assay_extract('taxonomy')
MAE |>
 EMP_assay_extract('geno_ko') 
##  Search for specific features according to the rowdata
MAE |>
  EMP_assay_extract('geno_ec',pattern = '1.1.1.1',pattern_ref = 'feature',exact = TRUE)
MAE |>
  EMP_assay_extract('geno_ko',pattern = 'mtlD',pattern_ref = 'Name',exact = FALSE)
}
