% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{EMP_boxplot}
\alias{EMP_boxplot}
\alias{EMP_boxplot,EMP_alpha_analysis-method}
\alias{EMP_boxplot,EMP_assay_data-method}
\title{Title}
\usage{
EMP_boxplot(obj, ...)

\S4method{EMP_boxplot}{EMP_alpha_analysis}(obj, ...)

\S4method{EMP_boxplot}{EMP_assay_data}(obj, ...)
}
\arguments{
\item{obj}{obj}

\item{...}{...}
}
\value{
xx object

xx object

xx object
}
\description{
Title

Title

Title
}
\examples{
# xx
#
#
}
