% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_alpha_boxplot.R,
%   R/Plot_EMP_boxplot_assay.R
\name{EMP_boxplot}
\alias{EMP_boxplot}
\alias{EMP_boxplot,EMP_alpha_analysis-method}
\alias{EMP_boxplot,EMP_assay_boxplot_union-method}
\alias{EMP_boxplot.EMP_alpha_analysis}
\alias{EMP_boxplot.EMP_assay_boxplot_union}
\title{Boxplot for EMPT result}
\usage{
EMP_boxplot(obj, ...)

\S4method{EMP_boxplot}{EMP_alpha_analysis}(obj, ...)

\S4method{EMP_boxplot}{EMP_assay_boxplot_union}(obj, ...)

EMP_boxplot.EMP_alpha_analysis(
  obj,
  plot_category = "default",
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  dot_size = 2,
  box_width = if (plot_category == "violin") 0.3 else NULL,
  box_alpha = if (plot_category == "violin") 0.8 else 1,
  step_increase = 0.1,
  ref.group = NULL,
  comparisons = NULL,
  ncol = NULL,
  select_metrics = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()",
  ...
)

EMP_boxplot.EMP_assay_boxplot_union(
  obj,
  plot_category = 1,
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  dot_size = 2,
  box_width = if (plot_category == "violin") 0.3 else NULL,
  box_alpha = if (plot_category == "violin") 0.8 else 1,
  step_increase = 0.1,
  ref.group = NULL,
  comparisons = NULL,
  ncol = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()",
  ...
)
}
\arguments{
\item{obj}{EMPT object}

\item{...}{Additional parameters, see also \code{\link[ggsignif]{geom_signif}}}

\item{plot_category}{A character string including default and violin.}

\item{method}{A character string. The name of the statistical test that is applied to the values of the columns (e.g. t.test, wilcox.test etc.).}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{dot_size}{A numeric. Set the dot size.}

\item{box_width}{A numeric. Set the box width.}

\item{box_alpha}{A numeric. Set the box alpha.}

\item{step_increase}{A numeric vector with the increase in fraction of total height for every additional comparison to minimize overlap.}

\item{ref.group}{a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of 2 values on the x-axis or the 2 integers that correspond to the index of the columns of interest.(default:NULL)}

\item{ncol}{An interger. Set the col number in the facet plot.}

\item{select_metrics}{A series of character string. Select the alpha metrics to show.}

\item{show}{A character string include pic (default), html.}

\item{palette}{A series of character string. Color palette.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
EMPT object
}
\description{
Boxplot for EMPT result
}
\examples{
data(MAE)
## from assay
MAE |> 
  EMP_assay_extract('host_gene',pattern = 'A1BG',
        pattern_ref = 'feature') |>
  EMP_boxplot(method='t.test',estimate_group='Group')

## from alpha analysis
MAE |> 
  EMP_assay_extract('taxonomy') |> 
  EMP_alpha_analysis()|>
  EMP_boxplot(method='t.test',estimate_group='Group')

# volin style and more parameter
MAE |> 
  EMP_assay_extract('host_gene',
                    pattern = 'A1BG',pattern_ref = 'feature') |>
  EMP_boxplot(estimate_group='Group',box_alpha=0.8,
              box_width=0.3,dot_size=5,plot_category='violin')
 MAE |>
   EMP_assay_extract(experiment='taxonomy')|> 
   EMP_alpha_analysis() |>
   EMP_boxplot(estimate_group='Group',method='t.test',
               box_alpha=0.8,dot_size=3,
               box_width=0.3,plot_category='violin')
}
