% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_alpha_boxplot.R,
%   R/Plot_EMP_boxplot_assay.R
\name{EMP_boxplot}
\alias{EMP_boxplot}
\alias{EMP_boxplot,EMP_alpha_analysis-method}
\alias{EMP_boxplot,EMP_assay_data-method}
\alias{EMP_boxplot,EMP_decostand-method}
\alias{EMP_boxplot.EMP_alpha_analysis}
\alias{EMP_boxplot.EMP_assay_data}
\title{Boxplot for EMPT result}
\usage{
EMP_boxplot(obj, ...)

\S4method{EMP_boxplot}{EMP_alpha_analysis}(obj, ...)

\S4method{EMP_boxplot}{EMP_assay_data}(obj, ...)

\S4method{EMP_boxplot}{EMP_decostand}(obj, ...)

EMP_boxplot.EMP_alpha_analysis(
  EMPT,
  plot_category = 1,
  seed = 123,
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  ncol = NULL,
  select_metrics = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)

EMP_boxplot.EMP_assay_data(
  EMPT,
  plot_category = 1,
  seed = 123,
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  ncol = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{obj}{EMPT object}

\item{...}{...}

\item{EMPT}{Object in EMPT format.}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{seed}{An interger. Set the random seed to the plot.}

\item{method}{A character string. The name of the statistical test that is applied to the values of the columns (e.g. t.test, wilcox.test etc.).}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{ncol}{An interger. Set the col number in the facet plot.}

\item{select_metrics}{A series of character string. Select the alpha metrics to show.}

\item{show}{A character string include pic (default), html.}

\item{palette}{A series of character string. Color palette.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
EMPT object
}
\description{
Boxplot for EMPT result
}
\examples{
data(MAE)
## from assay
MAE |> 
  EMP_assay_extract('host_gene',pattern = 'A1BG',pattern_ref = 'feature') |>
  EMP_boxplot(method='t.test',estimate_group='Group')

## from alpha analysis
MAE |> 
  EMP_assay_extract('taxonomy') |> 
  EMP_alpha_analysis()|>
  EMP_boxplot(method='t.test',estimate_group='Group')
# add example
# add example
}
