% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_alpha_boxplot.R
\name{EMP_boxplot_alpha}
\alias{EMP_boxplot_alpha}
\title{Title}
\usage{
EMP_boxplot_alpha(EMPT, plot_category = 1, seed = 123, ...)
}
\arguments{
\item{EMPT}{Object in EMPT or MultiAssayExperiment format.}

\item{plot_category}{An interger.More plot style.}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{...}{Other parameters:
method: The name of the statistical test that is applied to the values of the 2 columns (e.g. t.test, wilcox.test etc.).
estimate_group: A character string. Select the colname in the coldata to compare the data in the statistical test.
group_level: A string vector. Set the group order in the plot.
ncol: An interger. Set the col number in the facet plot.
select_metrics: A string vector. Select the alpha metrics you want to plot.
palette: A character string.Color palette.
show: A character string include pic (default), html.
html_width: An interger. Set the html width.
html_height: An interger. Set the html height.
mytheme: Modify components of a theme according to the ggplot2::theme.}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
