% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_boxplot_assay.R
\name{EMP_boxplot_assay_default}
\alias{EMP_boxplot_assay_default}
\title{Title}
\usage{
EMP_boxplot_assay_default(
  EMPT,
  method = "wilcox.test",
  estimate_group = NULL,
  group_level = "default",
  ncol = NULL,
  show = "pic",
  palette = NULL,
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{method}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{group_level}{wait_for_add}

\item{ncol}{wait_for_add}

\item{show}{wait_for_add}

\item{palette}{wait_for_add}

\item{html_width}{wait_for_add}

\item{html_height}{wait_for_add}

\item{mytheme}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
