% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_cluster.R
\name{EMP_cluster_analysis}
\alias{EMP_cluster_analysis}
\title{Identify clusters for sample or feature}
\usage{
EMP_cluster_analysis(
  x,
  experiment,
  distance = "bray",
  rowdata = FALSE,
  method = "average",
  h = NULL,
  groupLabels = TRUE,
  action = "add"
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{distance}{A character string.Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq", "chord", "hellinger", "aitchison", or "robust.aitchison".}

\item{rowdata}{A boolean. Whether the function cluster the feature or not.}

\item{method}{The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{h}{A numeric. Height at which to cut tree (passed to cutree)}

\item{groupLabels}{A boolean. Whether show the group label or not.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Identify clusters for sample or feature
}
\examples{
# add example
}
