% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_extract.R
\name{EMP_coldata_extract}
\alias{EMP_coldata_extract}
\title{Title}
\usage{
EMP_coldata_extract(
  obj,
  experiment = NULL,
  coldata_to_assay = NULL,
  assay_name = "undefined",
  action = "get"
)
}
\arguments{
\item{obj}{wait_for_add}

\item{experiment}{wait_for_add}

\item{coldata_to_assay}{wait_for_add}

\item{assay_name}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
