% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_collapse.R
\name{EMP_collapse}
\alias{EMP_collapse}
\title{Aggregates abundace or experssion from the same attributes in the coldata or rowdata}
\usage{
EMP_collapse(
  x,
  experiment,
  estimate_group = NULL,
  method = "sum",
  na_string = c("NA", "null", ""),
  collapse_by,
  collapse_sep = " ",
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{estimate_group}{A character string. Select the column in the rowdata or coldata to collapse.}

\item{method}{A character string. Methods include mean, sum, median, min, max.}

\item{na_string}{A series of character strings. Indicate which characters can be considered missing values.}

\item{collapse_by}{A character string. Methods include col or row.}

\item{collapse_sep}{A character string. The linking symbol used when strings are combined.}

\item{action}{A character string. A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function mean, sum, median, min, max in the base package.}
}
\value{
EMPT object
}
\description{
Aggregates abundace or experssion from the same attributes in the coldata or rowdata
}
\examples{
# add example
}
