% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_cor.R
\name{EMP_cor_analysis}
\alias{EMP_cor_analysis}
\title{Correlation analysis}
\usage{
EMP_cor_analysis(
  EMP,
  select = NULL,
  method = "spearman",
  action = "add",
  use_cached = TRUE,
  ...
)
}
\arguments{
\item{EMP}{Object in EMP format.}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string. Methods include pearson (default), spearman and kendall.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{...}{Further parameters passed to the function agricolae::correlation}
}
\value{
EMPT object
}
\description{
Correlation analysis
}
\examples{
data(MAE)
k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05 & abs(fold_change) > 1.5)

(k1 + k2) |> EMP_cor_analysis(method = 'spearman') |>
  EMP_heatmap_plot() ## Visualization
}
