% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_cor_analysis.R
\name{EMP_cor_analysis}
\alias{EMP_cor_analysis}
\title{Correlation analysis}
\usage{
EMP_cor_analysis(
  EMP,
  select = NULL,
  method = "spearman",
  action = "add",
  use_cached = T,
  ...
)
}
\arguments{
\item{EMP}{Object in EMP format.}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string. Methods include pearson (default), spearman and kendall.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{...}{Further parameters passed to the function agricolae::correlation}
}
\value{
EMPT object
}
\description{
Correlation analysis
}
\examples{
# add example
}
