% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/Plot_EMP_curveplot_enrich.R
\name{EMP_curveplot}
\alias{EMP_curveplot}
\alias{EMP_curveplot,EMP_enrich_analysis-method}
\alias{EMP_curveplot_enrich}
\title{Curveplot for enrichment result}
\usage{
EMP_curveplot(obj, ...)

\S4method{EMP_curveplot}{EMP_enrich_analysis}(obj, ...)

EMP_curveplot_enrich(obj, plot_category = 1, seed = 123, geneSetID, ...)
}
\arguments{
\item{obj}{Object in EMPT format.}

\item{...}{Further parameters passed to the function enrichplot::gseaplot2}

\item{plot_category}{An interger.More plot style.}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{geneSetID}{geneSet ID}
}
\value{
EMPT object
}
\description{
Curveplot for enrichment result
}
\examples{
\dontrun{
data(MAE)
MAE |>
 EMP_GSEA_analysis(experiment = 'geno_ko',method='signal2Noise',
                   estimate_group = 'Group',
                   pvalueCutoff = 0.05,keyType = 'ko') |>
 EMP_curveplot(geneSetID='map00680')
}
# add example
}
