% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_curveplot_enrich.R
\name{EMP_curveplot_enrich}
\alias{EMP_curveplot_enrich}
\title{curveplot for enrichment result}
\usage{
EMP_curveplot_enrich(EMPT, plot_category = 1, seed = 123, geneSetID, ...)
}
\arguments{
\item{EMPT}{Object in EMPT format.}

\item{plot_category}{An interger.More plot style.}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{geneSetID}{geneSet ID}

\item{...}{Further parameters passed to the function enrichplot::gseaplot2}
}
\value{
EMPT object
}
\description{
curveplot for enrichment result
}
\examples{
# add example
}
