% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_decostand.R
\name{EMP_decostand}
\alias{EMP_decostand}
\title{Standardization Methods}
\usage{
EMP_decostand(
  x,
  experiment,
  method,
  bySample = T,
  logbase = 2,
  use_cached = T,
  pseudocount = 1e-07,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Standardization method inluding relative(Equivalent to total in the vegan::decostand), integer, log, clr, alr, aclr,and more. Details see vegan::decostand.}

\item{bySample}{A boolean. Whether the function decostand by the sample or feature.}

\item{logbase}{An interger. The logarithm base used in method = "log".(default=2)}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{pseudocount}{A number. The logarithm pseudocount used in method = "clr".(default=0.0000001)}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function vegan::decostand.}
}
\value{
EMPT object
}
\description{
Standardization Methods
}
\examples{
# add example
}
