% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_diff.R
\name{EMP_diff_analysis}
\alias{EMP_diff_analysis}
\title{Differential expression or abundance analysis}
\usage{
EMP_diff_analysis(
  x,
  experiment,
  .formula,
  method = "wilcox.test",
  p.adjust = "fdr",
  estimate_group = NULL,
  use_cached = TRUE,
  action = "add",
  group_level = NULL,
  core = NULL,
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{.formula}{A formula representing the desired linear model. If there is more than one factor, they should be in the order factor of interest + additional factors.}

\item{method}{A character string. Methods include t.test, wilcox_test, kruskal.test, oneway.test, edgeR_quasi_likelihood, edgeR_likelihood_ratio, edger_robust_likelihood_ratio, DESeq2, limma_voom, limma_voom_sample_weights}

\item{p.adjust}{A character string. Adjust P-values for Multiple Comparisons inluding fdr, holm, hochberg, hommel, bonferroni, BH, BY. (default:fdr)}

\item{estimate_group}{A character string. Select the group name in the coldata to be calculated.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{group_level}{A series of character strings. Determine the comparison order of groups.}

\item{core}{A number. Select the core numbers in the parallel compute to speed up the result.}

\item{...}{Further parameters passed to the function tidybulk::test_differential_abundance, or statistical function in the stats package.}
}
\value{
EMPT object
}
\description{
Differential expression or abundance analysis
}
\examples{
# add example
}
