% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_diff.R
\name{EMP_diff_analysis}
\alias{EMP_diff_analysis}
\title{Differential expression or abundance analysis}
\usage{
EMP_diff_analysis(
  obj,
  experiment,
  .formula,
  method = "wilcox.test",
  p.adjust = "fdr",
  estimate_group = NULL,
  use_cached = TRUE,
  action = "add",
  group_level = NULL,
  core = NULL,
  ...
)
}
\arguments{
\item{obj}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{.formula}{A formula representing the desired linear model. If there is more than one factor, they should be in the order factor of interest + additional factors.}

\item{method}{A character string. Methods include t.test, wilcox.test, kruskal.test, oneway.test, edgeR_quasi_likelihood, edgeR_likelihood_ratio, edger_robust_likelihood_ratio, DESeq2, limma_voom, limma_voom_sample_weights}

\item{p.adjust}{A character string. Adjust P-values for Multiple Comparisons inluding fdr, holm, hochberg, hommel, bonferroni, BH, BY. (default:fdr)}

\item{estimate_group}{A character string. Select the group name in the coldata to be calculated.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{group_level}{A series of character strings. Determine the comparison order of groups. when group_level activted,.formula should be ~0+[your interested group] for edger and limma.}

\item{core}{A number. Select the core numbers in the parallel compute to speed up the result.When no core is set and the number of features exceeds 2000, the default will use all CPU cores minus one.}

\item{...}{Further parameters passed to the function \code{\link[tidybulk]{test_differential_abundance}} or \code{\link[stats]{t.test}}, \code{\link[stats]{wilcox.test}}, \code{\link[stats]{kruskal.test}},\code{\link[stats]{oneway.test}}.}
}
\value{
EMPT object
}
\description{
Differential expression or abundance analysis
}
\section{Detaild about p.adjust}{

When using common transcriptome differential analysis methods, such as DESeq2, it’s important to note that the method internally filters genes before adjusting p-values. 
As a result, the adjusted p-values may differ from those obtained using unfiltered data. 
To ensure consistency, you can first apply gene filtering using the EMP_identify_assay function with the edgeR method. 
This approach ensures that the adjusted p-values remain consistent.
}

\examples{
data(MAE)
## t.test or wilcox.test
MAE |>
  EMP_decostand(experiment = 'taxonomy',method = 'relative',pseudocount=0.0001) |>
  EMP_diff_analysis(method = 't.test',estimate_group = 'Group',p.adjust = 'fdr')

MAE |>
  EMP_assay_extract(experiment = 'taxonomy') |>
  EMP_diff_analysis(method = 'wilcox.test',estimate_group = 'Group',p.adjust = 'BH')

## DESeq2
MAE |>
  EMP_decostand(experiment = 'geno_ec',method = 'integer') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  

MAE |>
  EMP_decostand(experiment = 'geno_ec',method = 'integer') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Region+Group)  ## Eliminate the batch_effect in DESeq2


### edgeR_quasi_likelihood
MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='edgeR_quasi_likelihood',
                    .formula = ~0+Group,group_level = c('Group_B','Group_A')) ## Set the comparison order.
}
