% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_dimension_analysis.R
\name{EMP_dimension_analysis}
\alias{EMP_dimension_analysis}
\title{Title}
\usage{
EMP_dimension_analysis(
  x,
  experiment,
  method = "pca",
  distance = NULL,
  estimate_group = NULL,
  action = "add"
)
}
\arguments{
\item{x}{wait_for_add}

\item{experiment}{wait_for_add}

\item{method}{wait_for_add}

\item{distance}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{action}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
