% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_import.R
\name{EMP_easy_import}
\alias{EMP_easy_import}
\title{Easily import data into SummariseExperiment or MultiAssayExperiment}
\usage{
EMP_easy_import(
  file = NULL,
  data = NULL,
  type,
  assay = "experiment",
  assay_name = NULL,
  sampleID = NULL,
  coldata = NULL,
  file_format = NULL,
  humann_format = FALSE,
  duplicate_feature = NULL,
  output = "MAE",
  sep = if (humann_format == "TRUE") "|" else ";"
)
}
\arguments{
\item{file}{A file path. The file should be deposited in txt format.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{type}{A character string. Methods include tax, ko, ec and normal.}

\item{assay}{A character string. Set the experiment name.(default:experiment)}

\item{assay_name}{A character string. Indicate what kind of result the data belongs to, such as counts, relative abundance, TPM, etc.}

\item{sampleID}{A series of character strings. This parameter helps identify the assay and rowdata only when type = 'normal'.}

\item{coldata}{A dataframe containing one column informantion at least.}

\item{file_format}{A string including biom and gzv.}

\item{humann_format}{A boolean. Whether the function import the data in the Metaphlan or Humann format.}

\item{duplicate_feature}{A boolean. Whether the feature exist the dupicated name.}

\item{output}{A character string. Set the output result in SE(SummariseExperiment) or MAE(MultiAssayExperiment) format.}

\item{sep}{The field separator character. Only activated when type =''tax. Values on each line of the file are separated by this character. (defacult:'|')}
}
\value{
SummmariseExperiment or MultiAssayExperiment object
}
\description{
Easily import data into SummariseExperiment or MultiAssayExperiment
}
\details{
Paramter file_format and humann_format help the function import data properly. Data in humann format is usually is generated from Metaphlan and Humann. Data in biom format is usually is generated from Qiime1. Data in qzv format is usually is generated from Qiime2.
}
\examples{
# example
}
