% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_enrich.R
\name{EMP_enrich_analysis}
\alias{EMP_enrich_analysis}
\title{KEGG enrichment for the KO, EC and compound}
\usage{
EMP_enrich_analysis(
  x,
  condition,
  minGSSize = 1,
  maxGSSize = 500,
  keyType = NULL,
  KEGG_Type = "KEGG",
  species = "all",
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{condition}{Expressions that return a logical value according to the result of EMP_diff_analysis. eg. pvalue < 0.05}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{Maximal size of genes annotated for testing.}

\item{keyType}{A character string. keyType include ko, ec, cpd, entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to clusterProfiler::compareCluster.}
}
\value{
EMPT object
}
\description{
KEGG enrichment for the KO, EC and compound
}
\examples{
# add example
}
