% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_enrich.R
\name{EMP_enrich_analysis}
\alias{EMP_enrich_analysis}
\title{KEGG enrichment for EMPT object}
\usage{
EMP_enrich_analysis(
  x,
  condition,
  minGSSize = 1,
  maxGSSize = 500,
  keyType = NULL,
  KEGG_Type = "KEGG",
  species = "all",
  action = "add",
  combineGroup = FALSE,
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{condition}{Expressions that return a logical value according to the result of EMP_diff_analysis. eg. pvalue < 0.05}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{Maximal size of genes annotated for testing.}

\item{keyType}{A character string. keyType include ko, ec, cpd, entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{combineGroup}{A boolean. Whether the function combine the enrichment or not.}

\item{...}{Further parameters passed to clusterProfiler::compareCluster or clusterProfiler::enricher.}
}
\value{
EMPT object
}
\description{
KEGG enrichment for EMPT object
}
\examples{
\dontrun{
data(MAE)
## Make the enrichment after EMP_diff_analysis
MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group) |>
  EMP_enrich_analysis(keyType ='ec',KEGG_Type = 'KEGG',pvalue<0.05,pvalueCutoff=1,species = 'all') 

## Make the enrichment after EMP_diff_analysis and Visualization
MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group) |>
  EMP_enrich_analysis(keyType ='ec',KEGG_Type = 'KEGG',pvalue<0.05,pvalueCutoff=1,species = 'all') |>
  EMP_dotplot()

MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group) |>
  EMP_enrich_analysis(keyType ='ec',KEGG_Type = 'KEGG',pvalue<0.05,pvalueCutoff=1,species = 'all') |>
  EMP_netplot()


## Transcriptomic data
MAE |>
  EMP_assay_extract(experiment = 'host_gene') |>
  EMP_feature_convert(from = 'symbol',to='entrezid',species='Human') |>
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group,p.adjust = 'fdr') |> 
  EMP_enrich_analysis(keyType ='entrezid',KEGG_Type = 'KEGG',pvalue<0.05,pvalueCutoff=0.05,species = 'hsa') |>
  EMP_dotplot()
}
}
