% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_feature_covert.R
\name{EMP_feature_convert}
\alias{EMP_feature_convert}
\title{Covert feature of microbial taxonomy, gene experssion or compund abundance}
\usage{
EMP_feature_convert(
  obj,
  experiment,
  method = "mean",
  from,
  to = NULL,
  add = NULL,
  species = "none",
  OrgDb = NULL,
  action = "add"
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{method}{A character string. Methods include mean, sum, median, min, max. When multiple annotations appear on features, merge activate.}

\item{from}{A character string. For metabolite include CAS,DTXSID,DTXCID,SID,CID,KEGG,ChEBI,HMDB,Drugbank. For gene include SYMBOL,ENSEMBL,ENTREZID.}

\item{to}{A character string. For metabolite include CAS,DTXSID,DTXCID,SID,CID,KEGG,ChEBI,HMDB,Drugbank. For gene include SYMBOL,ENSEMBL,ENTREZID.}

\item{add}{A character string. For microbiome include tax_single and tax_full. For disease include Human_disease and Mouse_disease based on SYMBOL,ENTREZID,ko and ec.}

\item{species}{A character string. Species includ Human,Mouse,Pig,Zebrafish. If converting feature from other species,please use OrgDb.}

\item{OrgDb}{Supported OrgDb listed in 'https://bioconductor.org/packages/release/BiocViews.html#___OrgDb'}

\item{action}{A character string. A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Covert feature of microbial taxonomy, gene experssion or compund abundance
}
\examples{
data(MAE)
## for gene ID convert
MAE |>
  EMP_feature_convert(experiment = 'host_gene',from = 'SYMBOL',to='ENTREZID',species = 'Human')

## The built-in database only supports Human, Mouse, Pig, Zebrafish
## Other species could utilize OrgDb to convert
\dontrun{
library(org.Hs.eg.db)
MAE |>
  EMP_feature_convert(experiment = 'host_gene',from = 'SYMBOL',to='ENTREZID',OrgDb = org.Hs.eg.db)
}
## for compound ID convert
MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |> ## get the compound in KEGG format
  EMP_feature_convert(from = 'KEGG',to='HMDB')
## Add full name for microbial data
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_feature_convert(from = 'tax_single',add = 'tax_full') |>
  EMP_collapse(estimate_group = 'Phylum',collapse_by = 'row')
## Add disease info and select the related feature
MAE |> 
  EMP_assay_extract('host_gene') |>
  EMP_feature_convert(from = 'SYMBOL',add ='Human_disease') |>
  EMP_assay_extract(pattern = 'cancer',pattern_ref = 'Human_disease')
}
