% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_filter.R
\name{EMP_filter}
\alias{EMP_filter}
\title{Filer experssion or abudance data that match a condition}
\usage{
EMP_filter(
  obj,
  sample_condition,
  feature_condition,
  filterSample = NULL,
  filterFeature = NULL,
  experiment = NULL,
  show_info = NULL,
  action = "select"
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{sample_condition}{Expressions that return a logical value, and are defined in terms of the variables in coldata. If multiple expressions are included, they are combined with the &，| operator.}

\item{feature_condition}{Expressions that return a logical value, and are defined in terms of the variables in rowdata. If multiple expressions are included, they are combined with the &，| operator.}

\item{filterSample}{A series of character strings. Select samples in the data exactly.}

\item{filterFeature}{A series of character strings. Select samples in the data exactly.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{show_info}{A character string. Set the class of EMPT to show properly.}

\item{action}{A character string. You can use the filterSample and filterFeature parameters in conjunction with this. The choice is whether to keep filterSample and filterFeature (select), or simply exclude them (kick).}
}
\value{
EMPT object
}
\description{
Filer experssion or abudance data that match a condition
}
\examples{
# add example
}
