% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_import.R
\name{EMP_function_import}
\alias{EMP_function_import}
\title{Import kegg data into SummariseExperiment}
\usage{
EMP_function_import(
  file = NULL,
  data = NULL,
  type,
  assay_name = NULL,
  humann_format = FALSE
)
}
\arguments{
\item{file}{A file path. The file should be deposited in txt format.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{type}{A character string including ko and ec.}

\item{assay_name}{A character string. Indicate what kind of result the data belongs to, such as counts, relative abundance, TPM, etc.}

\item{humann_format}{A boolean. Whether the function improt the data according to the humann format.}
}
\value{
SummmariseExperiment object
}
\description{
Import kegg data into SummariseExperiment
}
\examples{
# More examples and tutorial could be found on website: 
# https://liubingdong.github.io/EasyMultiProfiler_tutorial/
}
