% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_GSEA.R
\name{EMP_GSEA_analysis}
\alias{EMP_GSEA_analysis}
\title{Title}
\usage{
EMP_GSEA_analysis(
  x,
  condition,
  experiment,
  estimate_group = NULL,
  method,
  cor_method = "pearson",
  group_level = NULL,
  type,
  use_cached = T,
  pseudocount = 1e-04,
  pvalueCutoff = 1,
  threshold = NULL,
  threshold_r = 0,
  threshold_p = 0.05,
  seed = T,
  action = "add",
  ...
)
}
\arguments{
\item{x}{wait_for_add}

\item{condition}{wait_for_add}

\item{experiment}{wait_for_add}

\item{estimate_group}{wait_for_add}

\item{method}{wait_for_add}

\item{cor_method}{wait_for_add}

\item{group_level}{wait_for_add}

\item{type}{wait_for_add}

\item{use_cached}{wait_for_add}

\item{pseudocount}{wait_for_add}

\item{pvalueCutoff}{wait_for_add}

\item{threshold}{wait_for_add}

\item{threshold_r}{wait_for_add}

\item{threshold_p}{wait_for_add}

\item{seed}{wait_for_add}

\item{action}{wait_for_add}

\item{...}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
