% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.EMP.utilis.R, R/Plot_EMP_heatmap.R
\name{EMP_heatmap_plot}
\alias{EMP_heatmap_plot}
\alias{EMP_heatmap_plot,EMP_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis2-method}
\alias{EMP_heatmap.EMP_cor_analysis}
\alias{EMP_heatmap.WGCNA}
\title{EMP heatmap plot}
\usage{
EMP_heatmap_plot(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_cor_analysis}(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis}(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis2}(obj, ...)

EMP_heatmap.EMP_cor_analysis(
  obj,
  palette = c("steelblue", "white", "darkred"),
  show = "all",
  mytheme = "theme()"
)

EMP_heatmap.WGCNA(
  obj,
  palette = c("steelblue", "white", "darkred"),
  show = "all",
  mytheme = "theme()"
)
}
\arguments{
\item{obj}{EMPT or EMP object}

\item{...}{...}

\item{palette}{1-3 character string. Color palette. (default: steelblue, white, darkred)}

\item{show}{A character string. Show inluding all, sig and pvalue.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
EMPT or EMP object
}
\description{
EMP heatmap plot
}
\examples{
\dontrun{
data(MAE)
## for cor analysis
k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05 & abs(fold_change) > 1.5)

(k1 + k2) \%>\% EMP_cor_analysis(method = 'spearman') |>
  EMP_heatmap_plot() ## Visualization
## for WGCNA
MAE |>
  EMP_assay_extract('geno_ec')  |> 
  EMP_identify_assay(method = 'edgeR',estimate_group = 'Group') |>
  EMP_WGCNA_cluster_analysis(RsquaredCut = 0.85,mergeCutHeight=0.4)  |>
  EMP_WGCNA_cor_analysis(coldata_to_assay = c('BMI','PHQ9','GAD7','HAMD','SAS','SDS'),method='spearman') |>
  EMP_heatmap_plot(palette = 'Spectral') 
}
}
