% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.EMP.utilis.R, R/EMP_heatmap.R
\name{EMP_heatmap_plot}
\alias{EMP_heatmap_plot}
\alias{EMP_heatmap_plot,EMP_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis2-method}
\alias{EMP_heatmap_cor}
\alias{EMP_heatmap_WGCNA}
\title{EMP heatmap plot}
\usage{
EMP_heatmap_plot(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_cor_analysis}(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis}(obj, ...)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis2}(obj, ...)

EMP_heatmap_cor(
  obj,
  palette = c("steelblue", "white", "darkred"),
  show = "all",
  mytheme = "theme()"
)

EMP_heatmap_WGCNA(
  obj,
  palette = c("steelblue", "white", "darkred"),
  show = "all",
  mytheme = "theme()"
)
}
\arguments{
\item{obj}{EMPT or EMP object}

\item{...}{...}

\item{palette}{1-3 character string. Color palette. (default: steelblue, white, darkred)}

\item{show}{A character string. Show inluding all, sig and pvalue.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
EMPT or EMP object
}
\description{
EMP heatmap plot
}
\examples{
#x
# add example
# add example
}
