% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.EMP.utilis.R
\name{EMP_heatmap_plot}
\alias{EMP_heatmap_plot}
\alias{EMP_heatmap_plot,EMP_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis-method}
\alias{EMP_heatmap_plot,EMP_WGCNA_cor_analysis2-method}
\title{Title}
\usage{
EMP_heatmap_plot(obj)

\S4method{EMP_heatmap_plot}{EMP_cor_analysis}(obj)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis}(obj)

\S4method{EMP_heatmap_plot}{EMP_WGCNA_cor_analysis2}(obj)
}
\arguments{
\item{obj}{x}

\item{EMP_cor_analysis}{x}

\item{EMP_WGCNA_cor_analysis}{x}

\item{EMP_WGCNA_cor_analysis2}{x}
}
\value{
xx object

xx object

xx object

x object
}
\description{
Title

Title

Title

Title
}
\examples{
#x
# x
# x
# 
}
