% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_history}
\alias{EMP_history}
\title{Get history from EMPT or EMP object}
\usage{
EMP_history(obj)
}
\arguments{
\item{obj}{Object in EMPT or EMP format.}
}
\value{
EMPT object
}
\description{
Get history from EMPT or EMP object
}
\examples{
data(MAE)
## from EMPT
MAE |>
  EMP_assay_extract(experiment = 'taxonomy') |>
  EMP_alpha_analysis() |>
  EMP_history()

## from EMP
k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05 & abs(fold_change) > 1.5)

(k1 + k2) |> EMP_cor_analysis(method = 'spearman') |>
  EMP_heatmap_plot()  |>
  EMP_history()  ## get all the history about what happened to the object
}
