% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_identify_assay.R
\name{EMP_identify_assay}
\alias{EMP_identify_assay}
\title{Identify the most core experssion and abudnace from sparse data}
\usage{
EMP_identify_assay(
  x,
  experiment,
  estimate_group = NULL,
  method = c("default", "edgeR"),
  min = if (method == "edgeR") 10 else 0.001,
  min_ratio = 0.7,
  action = "add"
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{estimate_group}{A character string. Select the group name in the coldata to be calculated.}

\item{method}{A character string.Methods include default, edgeR. Method default is from doi: 10.3389/fgene.2021.803627. Method edgeR in from edgeR::filterByExpr.}

\item{min}{A number. Set the min abundance for filtering. When method='default', min means the lowest relative bundance. When method='edgeR.', min means the lowest abosulte bundance.}

\item{min_ratio}{Set the min ratio presence for feature.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Identify the most core experssion and abudnace from sparse data
}
\examples{
# add example
}
