% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_identify_assay.R
\name{EMP_identify_assay}
\alias{EMP_identify_assay}
\title{Identify the most core experssion and abudnace from sparse data}
\usage{
EMP_identify_assay(
  obj,
  experiment = NULL,
  estimate_group = NULL,
  method = c("default", "edgeR"),
  min = if (method == "edgeR") 10 else 0.001,
  min_ratio = 0.7,
  use_cached = TRUE,
  action = "add"
)
}
\arguments{
\item{obj}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{estimate_group}{A character string. Select the group name in the coldata to be calculated. When estimate_group = NULL or "none", the function will assume all samples belong to one group.}

\item{method}{A character string.Methods include default, edgeR. Method default is from doi: 10.3389/fgene.2021.803627. Method edgeR in from \code{\link[edgeR]{filterByExpr}}.}

\item{min}{A number. Set the min abundance for filtering. When method='default', min means the lowest relative bundance. When method='edgeR.', min means the lowest abosulte bundance.}

\item{min_ratio}{Set the min ratio presence for feature.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Identify the most core experssion and abudnace from sparse data
}
\examples{
data(MAE)
## Consider the minnum relative abundance and min ratio specilally for microbial data
## First, the abundance below this threshold will be converted to 0 according to the set minimum species relative abundance. Finally, the core species will be required to meet the requirement that the occurrence rate of the species group in at least one group is higher than the preset threshold, and the remaining species will Then it will be judged as "rare species" and filtered.
## Note: If absolute abundance is provided as input, it will be automatically converted to relative abundance for filtering purposes during calculations. However, the output will remain in absolute abundance.
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_identify_assay(estimate_group = 'Group',method = 'default',
                     min=0.01,min_ratio = 0.7) # consider the Group
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_identify_assay(method = 'default') # consider all samples belong to one group

## Consider the minnum counts abundance and min ratio specilally for microbial data
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_identify_assay(method = 'edgeR',min = 10,min_ratio = 0.7) # consider all samples belong to one group

}
