% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_impute.R
\name{EMP_impute}
\alias{EMP_impute}
\title{Fast Imputation of Missing Values by Chained Random Forests}
\usage{
EMP_impute(
  x,
  experiment,
  coldata = TRUE,
  assay = FALSE,
  rowdata = FALSE,
  .formula = . ~ .,
  pmm.k = 10,
  num.trees = 1000,
  seed = 123,
  verbose = 0,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{coldata}{A boolean. Whether the function use the coldata to impute or not.(default:TRUE)}

\item{assay}{A boolean. Whether the function use the assay to impute or not.(default:FALSE)}

\item{rowdata}{A boolean. Whether the function use the rowdata to impute or not.(default:FALSE)}

\item{.formula}{A formula.A two-sided formula specifying variables to be imputed (left hand side) and variables used to impute (right hand side). Defaults to . ~ ., i.e., use all variables to impute all variables. For instance, if all variables (with missings) should be imputed by all variables except variable "ID", use . ~ . - ID. Note that a "." is evaluated separately for each side of the formula. Further note that variables with missings must appear in the left hand side if they should be used on the right hand side.}

\item{pmm.k}{Number of candidate non-missing values to sample from in the predictive mean matching steps. 0 to avoid this step.}

\item{num.trees}{Tree number.}

\item{seed}{Integer seed to initialize the random generator.}

\item{verbose}{Controls how much info is printed to screen. 0 to print nothing.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function missRanger::missRanger.}
}
\value{
EMPT object
}
\description{
Fast Imputation of Missing Values by Chained Random Forests
}
\examples{
\dontrun{
data(MAE)
## For coldata
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_impute(assay=FALSE,coldata=TRUE,rowdata=FALSE) |>
  EMP_coldata_extract()
## Support formula, such as only impute SAS and SDS
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_impute(.formula = HAMA+HAMD ~ .) |>
  EMP_coldata_extract()

## For assay
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_impute(assay=TRUE,coldata=FALSE,rowdata=FALSE)

## For rowdata (Not Not recommended)
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_impute(assay=FALSE,coldata=FALSE,rowdata=TRUE)
}
}
