% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_impute.R
\name{EMP_impute}
\alias{EMP_impute}
\title{Fast Imputation of Missing Values by Chained Random Forests}
\usage{
EMP_impute(
  x,
  experiment,
  coldata = TRUE,
  assay = FALSE,
  rowdata = FALSE,
  pmm.k = 10,
  num.trees = 1000,
  seed = 123,
  verbose = 0,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{coldata}{A boolean. Whether the function use the coldata to impute or not.(default:TRUE)}

\item{assay}{A boolean. Whether the function use the assay to impute or not.(default:FALSE)}

\item{rowdata}{A boolean. Whether the function use the rowdata to impute or not.(default:FALSE)}

\item{pmm.k}{Number of candidate non-missing values to sample from in the predictive mean matching steps. 0 to avoid this step.}

\item{num.trees}{Tree number.}

\item{seed}{Integer seed to initialize the random generator.}

\item{verbose}{Controls how much info is printed to screen. 0 to print nothing.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function missRanger::missRanger.}
}
\value{
EMPT object
}
\description{
Fast Imputation of Missing Values by Chained Random Forests
}
\examples{
# add example
}
