% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_modify_assay.R
\name{EMP_modify_assay}
\alias{EMP_modify_assay}
\title{Modify assay in condition for some special cases}
\usage{
EMP_modify_assay(
  obj,
  condition = "==0",
  experiment,
  select_sample = "all",
  select_feature = "all",
  pseudocount = 1e-04,
  action = "add"
)
}
\arguments{
\item{obj}{MAE or EMPT objetc}

\item{condition}{A character string. Set the condition data will modify.(default: '==0')}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{select_sample}{A character string. Choose which sample to modify.}

\item{select_feature}{A character string. Choose which feature to modify.}

\item{pseudocount}{A number. Set the pseudocount data will change into. (default: '0.0001')}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMPT object
}
\description{
Modify assay in condition for some special cases
}
\examples{
data(MAE)
## For some special cases, to modify the assay data and EMP_identify_assay works better in most cases.
## Change the expression value which is 0 into 0.0001
MAE |>
  EMP_assay_extract('geno_ec') |>
  EMP_modify_assay('==0',pseudocount=0.0001)

## Change the counts which is below 
MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_modify_assay('<10',pseudocount=0) 
}
