% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_multi.R
\name{EMP_multi_analysis}
\alias{EMP_multi_analysis}
\title{Multi analysis for diff reslut or enrichment}
\usage{
EMP_multi_analysis(
  EMP,
  select = NULL,
  method = "feature",
  combineFun = "enricher",
  pvalueCutoff = 0.05,
  combineMethod = "fisher",
  combineGroup = FALSE,
  combineLevel = "gene",
  p.adjust = "fdr",
  minGSSize = 10,
  maxGSSize = 500,
  keyType = "ko",
  KEGG_Type = "KEGG",
  species = "all",
  action = "add",
  ...
)
}
\arguments{
\item{EMP}{EMP objetc}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string including feature, diff_feature_enrich and same_feature_enrich.}

\item{combineFun}{A character string including enricher, ActivePathways and mitch.}

\item{pvalueCutoff}{A character string. Adjusted pvalue cutoff on enrichment tests to report.}

\item{combineMethod}{A character string including fisher, edgington and stouffer. Only actived, when combineFun == "enricher".}

\item{combineGroup}{A boolean. Whether the function combine the enrichment or not.}

\item{combineLevel}{A character string. One of "gene" and "enrichResult". Only actived, when combineFun == "enricher".}

\item{p.adjust}{A character string. One of holm, hochberg, hommel, bonferroni, BH, BY, fdr, none.(defaault: fdr)}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing. (default=10)}

\item{maxGSSize}{Maximal size of each geneSet for analyzing. default=500)}

\item{keyType}{A character string. Methods include ko, ec, cpd and entrezid.}

\item{KEGG_Type}{A character string. KEGG_Type include KEGG and MKEGG.}

\item{species}{A character string. Species includ all, hsa, mmu,...Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{...}
}
\value{
EMPT object
}
\description{
Multi analysis for diff reslut or enrichment
}
\examples{
# add example
}
