% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R, R/EMP_enrich_netplot.R
\name{EMP_netplot}
\alias{EMP_netplot}
\alias{EMP_netplot,EMP_enrich_analysis-method}
\alias{EMP_netplot_enrich}
\title{Title}
\usage{
EMP_netplot(obj, ...)

\S4method{EMP_netplot}{EMP_enrich_analysis}(obj, ...)

EMP_netplot_enrich(
  EMPT,
  plot_category = 1,
  seed = 123,
  layout = "kk",
  showCategory = 5,
  ...
)
}
\arguments{
\item{obj}{object}

\item{...}{Further parameters passed to enrichplot::cnetplot.}

\item{EMPT}{EMPT object}

\item{plot_category}{wait_for_add}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{layout}{Layout of the map, e.g. 'star', 'circle', 'gem', 'dh', 'graphopt', 'grid', 'mds', 'randomly', 'fr', 'kk', 'drl' or 'lgl'.}

\item{showCategory}{A number or a vector of terms. If it is a number, the first n terms will be displayed. If it is a vector of terms, the selected terms will be displayed.}

\item{EMP_enrich_analysis}{xx}
}
\value{
EMP_enrich_analysis_netplot object

EMP_enrich_analysis_netplot object

EMPT object
}
\description{
Title

Title

Netplot for enrichment result
}
\examples{
# 
#
# add example
}
