% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyisis_EMP_network.R
\name{EMP_network_analysis}
\alias{EMP_network_analysis}
\title{Network analysis}
\usage{
EMP_network_analysis(
  obj,
  select = NULL,
  method = "cor",
  corMethod = "spearman",
  weighted = TRUE,
  missing = "pairwise",
  threshold = 0,
  coldata_to_assay = NULL,
  use_cached = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{obj}{EMPT or EMP.}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string. Methods include cor, EBICglasso,IsingFit,IsingSampler,pcor,huge,mgm,TMFG,LoGo,relimp,ggmModSelect and graphicalVAR.}

\item{corMethod}{A character string. Methods include spearman, cor, cov, npn and cor_auto.}

\item{weighted}{Logical, should the analyzed network be weighted?}

\item{missing}{How to handle missing data? "pairwise" for pairwise deletion, "listwise" for listwise deletion, "fiml" for full-information maximum likelihood and "stop" to stop with an error.}

\item{threshold}{Thresholding to use in partial correlation networks. Can be a fixed number to threshold all absolute edges below this value, 'locfdr' for local FDR, or any option corresponding to adjustments in \code{\link[psych]{corr.p}} ('none', 'sig', 'holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY' or 'fdr')). Can also be used for default = "IsingSampler" but can only be set to a logical enabling or disabling significance thresholding.}

\item{coldata_to_assay}{A series of character strings. Select the column from coldata to caculate. Only for the EMPT object.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function \code{\link[bootnet]{estimateNetwork}}.}
}
\value{
EMP object
}
\description{
Network analysis
}
\section{Detaild about network}{

More detailed information are available on :

Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Estimating psychological networks and their accuracy: A tutorial paper. Behavior Research Methods, 50(1), 195–212.
}

\examples{
data(MAE)

k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='wilcox.test', estimate_group = 'Group') |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)


#For single experiment
k1  |>
  EMP_network_analysis() |>
  EMP_filter(feature_condition = top_detect(Closeness,1))  # filter the top node according to the network analysis

#For muti experiment
(k1 + k2 ) |> 
  EMP_network_analysis() |> 
  EMP_network_plot(show = 'node') # get the node importance
}
