% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyisis_EMP_network.R
\name{EMP_network_analysis}
\alias{EMP_network_analysis}
\title{Network analysis}
\usage{
EMP_network_analysis(
  obj,
  select = NULL,
  method = "cor",
  corMethod = "spearman",
  weighted = TRUE,
  missing = "listwise",
  threshold = "sig",
  coldata_to_assay = NULL,
  use_cached = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{obj}{EMPT or EMP.}

\item{select}{A character string. The experiment name in the EMP object.}

\item{method}{A character string. Methods include cor, EBICglasso,IsingFit,IsingSampler,pcor,huge and graphicalVAR.}

\item{corMethod}{A character string. Methods include spearman, cor, cov, npn and cor_auto.}

\item{weighted}{Logical, should the analyzed network be weighted?}

\item{missing}{How to handle missing data? "pairwise" for pairwise deletion, "listwise" for listwise deletion, "fiml" for full-information maximum likelihood and "stop" to stop with an error.}

\item{threshold}{This parameter allows input of strings such as 'none', 'sig', 'holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', or 'fdr' for statistical calculation on edges, with the default set to 'sig' to select only statistically significant edges; alternatively, it can be a numeric value between 0 and 1, which will remove edges with an absolute value greater than this threshold.}

\item{coldata_to_assay}{A series of character strings. Select the column from coldata to caculate. Only for the EMPT object.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string.A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function \code{\link[bootnet]{estimateNetwork}}.}
}
\value{
EMP or EMPT object
}
\description{
Network analysis
}
\section{Detaild about network}{

More detailed information are available on :

Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Estimating psychological networks and their accuracy: A tutorial paper. Behavior Research Methods, 50(1), 195–212. 

Here is a brife introduction of document from bootnet package:

1. cor: Correlation network.

2. EBICglasso: Gaussian Markov random field estimation using graphical LASSO and extended Bayesian information criterion to select optimal regularization parameter. Using \code{\link[qgraph]{EBICglasso}} from the qgraph package. Calls bootnet_EBICglasso.

3. IsingSampler: Calls the \code{\link[IsingSampler]{EstimateIsing}} function from the IsingSampler package.

4. pcor: Partial correlation network (non-regularized Gaussian Markov random field), using \code{\link[corpcor]{cor2pcor}} from the corpcor package. Calls bootnet_pcor.

5. huge: Uses EBIC model selection of GGM networks estimated via the glasso algorithm as implemented in the huge package (as opposed to glasso and qgraph packages used in default = "EBICglasso"). Uses nonparanormal transformation in preparing the data and does not use polychoric correlations. Calls bootnet_huge.

6. graphicalVAR.: Estimates a graphical VAR model using the graphicalVAR package. This results in two networks which can be plotted using the 'graph' argument in the plot method. Calls bootnet_graphicalVAR.
}

\examples{
data(MAE)

k1 <- MAE |>
  EMP_assay_extract('taxonomy') |>
  EMP_collapse(estimate_group = 'Genus',collapse_by = 'row') |>
  EMP_diff_analysis(method='wilcox.test', estimate_group = 'Group') |>
  EMP_filter(feature_condition = pvalue<0.05)

k2 <- MAE |>
  EMP_collapse(experiment = 'untarget_metabol',na_string=c('NA','null','','-'),
               estimate_group = 'MS2kegg',method = 'sum',collapse_by = 'row') |>
  EMP_diff_analysis(method='DESeq2', .formula = ~Group) |>
  EMP_filter(feature_condition = pvalue<0.05)


#For single experiment
k1  |>
  EMP_network_analysis() |>
  EMP_filter(feature_condition = top_detect(Closeness,1))  # filter the top node according to the network analysis

#For muti experiment
(k1 + k2 ) |> 
  EMP_network_analysis() |> 
  EMP_network_plot(show = 'node') # get the node importance
}
