% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{EMP_normal_import}
\alias{EMP_normal_import}
\title{Title}
\usage{
EMP_normal_import(
  file = NULL,
  data = NULL,
  sampleID = NULL,
  dfmap = NULL,
  assay_name = NULL,
  assay = NULL
)
}
\arguments{
\item{file}{A file path.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{sampleID}{a seris of string character.}

\item{dfmap}{A dataframe. Indicate the experiment name, sample source and sample tube details in the omics data.}

\item{assay_name}{A character string. Indicate what kind of result the data belongs to, such as counts, relative abundance, TPM, etc.}

\item{assay}{A character string. Indicate the experiment name of the import data in the dfmap.}
}
\value{
SummmariseExperiment object
}
\description{
Title
}
\examples{
# example
}
