% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.EMPT.utilis.R
\name{EMP_result}
\alias{EMP_result}
\alias{EMP_result,EMPT-method}
\alias{EMP_result<-}
\alias{EMP_result<-,EMPT-method}
\title{Extract the existed result from EMPT or inject external result into EMPT}
\usage{
EMP_result(obj, info)

\S4method{EMP_result}{EMPT}(obj, info)

EMP_result(
  obj,
  value_name,
  affect_when_sample_changed,
  affect_when_feature_changed,
  attribute,
  attribute2,
  source
) <- value

\S4method{EMP_result}{EMPT}(
  obj,
  value_name,
  affect_when_sample_changed,
  affect_when_feature_changed,
  attribute,
  attribute2,
  source
) <- value
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{info}{A character string. Result or analysis name in the EMPT object.}

\item{value_name}{A character string. Set the name of external result to inject into EMPT project.}

\item{affect_when_sample_changed}{0 or 1. 0 means that the result is not influenced by sample changes, while 1 means the contrary.}

\item{affect_when_feature_changed}{0 or 1. 0 means that the result is not influenced by feature changes, while 1 means the contrary.}

\item{attribute}{A character string inculding primary, feature, all. This parameter indicates whether the result is about primary, feature or all.}

\item{attribute2}{A character string inculding normal, diagonal or none. This parameter indicates the format of result.}

\item{source}{A character string. Set the name of the analysis which generate the result. (default: user_import)}

\item{value}{A data frame or tibble from the external result.}
}
\value{
xx

xx
}
\description{
Extract the existed result from EMPT or inject external result into EMPT
}
\examples{
# xx
# xx
}
