% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_EMP_extract.R
\name{EMP_rowdata_extract}
\alias{EMP_rowdata_extract}
\title{Extract rowdata}
\usage{
EMP_rowdata_extract(
  obj,
  experiment = NULL,
  pattern_ref = "Name",
  pattern = NULL,
  exact = FALSE
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{pattern_ref}{A character string. Select which column in the rowdata to extract rowdata from.}

\item{pattern}{A character string. Select which pattern in the feature to extract rowdata.}

\item{exact}{A boolean. Whether the extract use exact search method.(default:FALSE)}
}
\value{
table
}
\description{
Extract rowdata
}
\examples{
data(MAE)
## Extract the rowdata of one existed experiment from MultiAssaayExperiment
MAE |>
  EMP_rowdata_extract('taxonomy')

MAE |>
  EMP_rowdata_extract('geno_ko')  
  
## Extract the rowdata of MultiAssaayExperiment
MAE |>
  EMP_rowdata_extract(experiment = NULL) -> total_row_data
dim(total_row_data)
}
