% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_extract.R
\name{EMP_rowdata_extract}
\alias{EMP_rowdata_extract}
\title{Title}
\usage{
EMP_rowdata_extract(
  obj,
  experiment = NULL,
  pattern_ref = "Name",
  pattern = NULL,
  exact = F
)
}
\arguments{
\item{obj}{MultiAssayExperiment object}

\item{experiment}{the choosed}

\item{pattern_ref}{wait_for_add}

\item{pattern}{wait_for_add}

\item{exact}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
