% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_extract.R
\name{EMP_rowdata_extract}
\alias{EMP_rowdata_extract}
\title{Extract rowdata}
\usage{
EMP_rowdata_extract(
  obj,
  experiment = NULL,
  pattern_ref = "Name",
  pattern = NULL,
  exact = FALSE
)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{pattern_ref}{A character string. Select which column in the rowdata to extract rowdata from.}

\item{pattern}{A character string. Select which pattern in the feature to extract rowdata.}

\item{exact}{A boolean. Whether the extract use exact search method.(default:FALSE)}
}
\value{
table
}
\description{
Extract rowdata
}
\examples{
# add example
}
