% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation_EMP_rrarefy.R
\name{EMP_rrarefy}
\alias{EMP_rrarefy}
\title{Rarefaction abundance or experssion Richness}
\usage{
EMP_rrarefy(
  x,
  experiment,
  raresize = NULL,
  seed = 123,
  only_show_depth = FALSE,
  use_cached = TRUE,
  action = "add",
  ...
)
}
\arguments{
\item{x}{Object in EMPT or MultiAssayExperiment format.}

\item{experiment}{A character string. Experiment name in the MultiAssayExperiment object.}

\item{raresize}{An interger. Subsample size for rarefying community.}

\item{seed}{An interger. Set the random seed to the plot.(default:123)}

\item{only_show_depth}{A boolean. Whether the function only show the depth or not.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}

\item{...}{Further parameters passed to the function vegan::rrarefy.}
}
\value{
EMPT object
}
\description{
Rarefaction abundance or experssion Richness
}
\examples{
# add example
}
