% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_EMP_function.R
\name{EMP_save_var}
\alias{EMP_save_var}
\title{Save result into environment}
\usage{
EMP_save_var(data, var_name, envir = .GlobalEnv, ...)
}
\arguments{
\item{data}{data}

\item{var_name}{a name to be assigned to data.}

\item{envir}{the \link{environment} to use.}

\item{...}{Addtional parameters, see also \code{\link[base]{assign}}.}
}
\value{
EMPT object
}
\description{
Save result into environment
}
\examples{
data(MAE)
MAE |> 
  EMP_assay_extract('host_gene',pattern = 'A1BG',pattern_ref = 'feature') |>
  EMP_save_var('temp_data') |>
  EMP_collapse(estimate_group = 'Group',collapse_by = 'col') |>
  EMP_heatmap_plot()

temp_data
}
