% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\name{EMP_scatterplot}
\alias{EMP_scatterplot}
\alias{EMP_scatterplot,EMP_dimension_analysis-method}
\title{Title}
\usage{
EMP_scatterplot(obj, ...)

\S4method{EMP_scatterplot}{EMP_dimension_analysis}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{...}
}
\value{
EMP_dimension_analysis_scatterplot object

EMP_dimension_analysis_scatterplot object
}
\description{
Title

Title
}
\examples{
# 
#
}
