% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R,
%   R/Plot_EMP_scatterplot_reduce_dimension.R
\name{EMP_scatterplot}
\alias{EMP_scatterplot}
\alias{EMP_scatterplot,EMP_dimension_analysis-method}
\alias{EMP_scatterplot.EMP_dimension_analysis}
\title{Scatterplot for EMPT result}
\usage{
EMP_scatterplot(obj, ...)

\S4method{EMP_scatterplot}{EMP_dimension_analysis}(obj, ...)

EMP_scatterplot.EMP_dimension_analysis(
  EMPT,
  seed = 123,
  group_level = "default",
  show = "p12",
  distance_for_adonis = NULL,
  force_adonis = FALSE,
  estimate_group = NULL,
  palette = NULL,
  method = "t.test",
  key_samples = NULL,
  ellipse = NULL,
  html_width = 15,
  html_height = 15
)
}
\arguments{
\item{obj}{object}

\item{...}{...}

\item{EMPT}{EMPT object}

\item{seed}{An interger. Set the random seed to the plot.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{show}{A character string include pic (default), html.}

\item{distance_for_adonis}{A character string.Set the distance for adonis. Detailed in the vegan::adonis.}

\item{force_adonis}{force the function run adnois analysis always.(default:FALSE)}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{palette}{A series of character string. Color palette.}

\item{method}{A character string. The name of the statistical test that is applied to the values of the columns (e.g. t.test, wilcox.test etc.).}

\item{key_samples}{A series of character string. To highlight your interested samples.}

\item{ellipse}{A number from 0 to 1. Set the ellipse in the plot.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}
}
\value{
EMPT object
}
\description{
Scatterplot for EMPT result
}
\examples{
# 
# add example
# 
}
