% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R,
%   R/Plot_EMP_scatterplot_reduce_dimension.R
\name{EMP_scatterplot}
\alias{EMP_scatterplot}
\alias{EMP_scatterplot,EMP_dimension_analysis-method}
\alias{EMP_scatterplot.EMP_dimension_analysis}
\title{Scatterplot for EMPT result}
\usage{
EMP_scatterplot(obj, ...)

\S4method{EMP_scatterplot}{EMP_dimension_analysis}(obj, ...)

EMP_scatterplot.EMP_dimension_analysis(
  obj,
  seed = 123,
  group_level = "default",
  show = "p12",
  distance_for_adonis = NULL,
  force_adonis = FALSE,
  adonis_permutations = 999,
  estimate_group = NULL,
  palette = NULL,
  method = "wilcox.test",
  key_samples = NULL,
  step_increase = 0.1,
  ref.group = NULL,
  comparisons = NULL,
  ellipse = NULL,
  html_width = 15,
  html_height = 15,
  ...
)
}
\arguments{
\item{obj}{EMPT object}

\item{...}{Additional parameters for adjust the boxplot, see also \code{\link[ggsignif]{geom_signif}}.}

\item{seed}{An interger. Set the random seed for the adonis permutations.}

\item{group_level}{A string vector. Set the group order in the plot.}

\item{show}{A character string include pic (default), html. This could display graphical results on 3 axes. eg. p12,p12html,p23,p23html}

\item{distance_for_adonis}{A character string.Set the distance for adonis. Detailed in \code{\link[vegan]{adonis}}.}

\item{force_adonis}{Force the function run adnois analysis always.(default:FALSE)}

\item{adonis_permutations}{Permutations for the adonis2.(default:999)}

\item{estimate_group}{A character string. Select the colname in the coldata to compare the data in the statistical test.}

\item{palette}{A series of character string. Color palette.}

\item{method}{A character string. The name of the statistical test that is applied to barplot columns (default:wilcox.test).}

\item{key_samples}{A series of character string. To highlight your interested samples.}

\item{step_increase}{A numeric vector with the increase in fraction of total height for every additional comparison to minimize overlap.}

\item{ref.group}{a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of 2 values on the x-axis or the 2 integers that correspond to the index of the columns of interest.(default:NULL)}

\item{ellipse}{A number from 0 to 1. Set the ellipse in the plot.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}
}
\value{
EMPT object
}
\description{
Scatterplot for EMPT result
}
\examples{
data(MAE)
MAE |> 
  EMP_assay_extract('taxonomy') |> 
  EMP_collapse(estimate_group = 'Species',collapse_by = 'row')|>
  EMP_dimension_analysis(method = 'pcoa',distance = 'bray')|>
  EMP_scatterplot(show='p12html',estimate_group = 'Group') # eg. p12,p12html,p23,p23html
}
