% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{EMP_taxonomy_import}
\alias{EMP_taxonomy_import}
\title{Title}
\usage{
EMP_taxonomy_import(
  x,
  humann_format = FALSE,
  assay_name = NULL,
  sep = if (humann_format == "FALSE") ";" else "|"
)
}
\arguments{
\item{x}{wait_for_add}

\item{humann_format}{wait_for_add}

\item{assay_name}{wait_for_add}

\item{sep}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
