% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_volcanol_plot.R
\name{EMP_volcanol_plot}
\alias{EMP_volcanol_plot}
\title{Title}
\usage{
EMP_volcanol_plot(
  EMPT,
  y = "pvalue",
  palette = NULL,
  show = "pic",
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{EMPT}{wait_for_add}

\item{y}{wait_for_add}

\item{palette}{wait_for_add}

\item{show}{wait_for_add}

\item{html_width}{wait_for_add}

\item{html_height}{wait_for_add}

\item{mytheme}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
