% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_EMP_volcanol.R
\name{EMP_volcanol_plot}
\alias{EMP_volcanol_plot}
\title{EMP volcanol plot}
\usage{
EMP_volcanol_plot(
  EMPT,
  y = "pvalue",
  palette = NULL,
  show = "pic",
  html_width = NULL,
  html_height = NULL,
  mytheme = "theme()"
)
}
\arguments{
\item{EMPT}{EMPT object}

\item{y}{A character string. Select the pvalue from the EMP_diff_analysis.}

\item{palette}{A series of character string. Color palette.}

\item{show}{A character string include pic (default), html.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}
}
\value{
EMPT object
}
\description{
EMP volcanol plot
}
\examples{
data(MAE)
MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  |>
  EMP_volcanol_plot(show='html')
}
