% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_EMP_volcanol.R
\name{EMP_volcanol_plot}
\alias{EMP_volcanol_plot}
\title{EMP volcanol plot}
\usage{
EMP_volcanol_plot(
  obj,
  plot_category = 1,
  y = "pvalue",
  palette = NULL,
  show = "pic",
  html_width = NULL,
  html_height = NULL,
  key_feature = NULL,
  mytheme = "theme()",
  ...
)
}
\arguments{
\item{obj}{EMPT object}

\item{plot_category}{An interger.More plot style.(under constrution)}

\item{y}{A character string. Select the pvalue from the EMP_diff_analysis.}

\item{palette}{A series of character string. Color palette.}

\item{show}{A character string include pic (default), html.}

\item{html_width}{An interger. Set the html width.}

\item{html_height}{An interger. Set the html height.}

\item{key_feature}{A series of character string. Label the some feature.}

\item{mytheme}{Modify components of a theme according to the ggplot2::theme.}

\item{...}{Further parameters passed to \code{\link[ggrepel]{geom_text_repel}}.}
}
\description{
EMP volcanol plot
}
\examples{
data(MAE)
MAE |>
  EMP_assay_extract(experiment = 'geno_ec') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  |>
  EMP_volcanol_plot(show='html')

# label feature
MAE |>
  EMP_decostand(experiment = 'geno_ec',method = 'integer') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  |>
  EMP_volcanol_plot(key_feature = c('3.6.1.62','1.5.3.19')) 

# Addtionl parameters will pass into ggrepel::geom_text_repel.
MAE |>
  EMP_decostand(experiment = 'geno_ec',method = 'integer') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  |>
  EMP_volcanol_plot(key_feature = c('3.6.1.62','1.5.3.19'),color = "white",
                    bg.color = "grey30",bg.r = 0.15)
MAE |>
  EMP_decostand(experiment = 'geno_ec',method = 'integer') |>
  EMP_diff_analysis(method='DESeq2',.formula = ~Group)  |>
  EMP_volcanol_plot(key_feature = c('3.6.1.62','1.5.3.19'),
                    min.segment.length = 0, seed = 42, box.padding = 0.5) ## Add arrow
}
