% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_EMP_WGCNA.R
\name{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis}
\title{EMP_WGCNA_cor_analysis}
\usage{
EMP_WGCNA_cor_analysis(
  obj,
  select = NULL,
  method = "pearson",
  coldata_to_assay = NULL,
  use_cached = TRUE,
  action = "add"
)
}
\arguments{
\item{obj}{EMPT or EMP object.}

\item{select}{A character string. The experiment name in the EMP object. Only for the EMP object.}

\item{method}{A character string. Methods include pearson, spearman and kendall.(default:pearson)}

\item{coldata_to_assay}{A series of character strings. Select the column from coldata to caculate. Only for the EMPT object.}

\item{use_cached}{A boolean. Whether the function use the results in cache or re-compute.}

\item{action}{A character string. Whether to join the new information to the EMPT (add), or just get the detailed result generated here (get).}
}
\value{
EMP object
}
\description{
EMP_WGCNA_cor_analysis
}
\examples{
\dontrun{
data(MAE)
## from one experiment
WGCNA_COR_result <- MAE |>
  EMP_assay_extract('geno_ec')  |> 
  EMP_identify_assay(method = 'edgeR',estimate_group = 'Group') |>
  EMP_WGCNA_cluster_analysis(RsquaredCut = 0.85,mergeCutHeight=0.4)  |>
  EMP_WGCNA_cor_analysis(coldata_to_assay = c('BMI','PHQ9','GAD7','HAMD','SAS','SDS'),
                         method='pearson',action='add') # If want the detailed result, set action = 'get'

## Visualization
MAE |>
  EMP_assay_extract('geno_ec')  |> 
  EMP_identify_assay(method = 'edgeR',estimate_group = 'Group') |>
  EMP_WGCNA_cluster_analysis(RsquaredCut = 0.85,mergeCutHeight=0.4)  |>
  EMP_WGCNA_cor_analysis(coldata_to_assay = c('BMI','PHQ9','GAD7','HAMD','SAS','SDS'),method='pearson') |>
  EMP_heatmap_plot(palette = 'Spectral')

## Filter the interesting module and make the enrichment analysis
MAE |>
  EMP_assay_extract('geno_ec')  |> 
  EMP_identify_assay(method = 'edgeR',estimate_group = 'Group') |>
  EMP_WGCNA_cluster_analysis(RsquaredCut = 0.85,mergeCutHeight=0.4)  |>
  EMP_WGCNA_cor_analysis(coldata_to_assay = c('BMI','PHQ9','GAD7','HAMD','SAS','SDS'),method='pearson') |>
  EMP_heatmap_plot(palette = 'Spectral') |>
  EMP_filter(feature_condition = WGCNA_color == 'brown' ) |> 
  EMP_diff_analysis(method = 'DESeq2',.formula = ~Group) |>
  EMP_enrich_analysis(keyType = 'ec',KEGG_Type = 'MKEGG') |>
  EMP_enrich_dotplot()

## from two different experiments
k1 <- MAE |>
  EMP_assay_extract('geno_ec')  |> 
  EMP_identify_assay(method = 'edgeR',estimate_group = 'Group') |>
  EMP_WGCNA_cluster_analysis(RsquaredCut = 0.85,mergeCutHeight=0.4)

k2 <- MAE |>
  EMP_assay_extract('host_gene',pattern = c('A1BG','A1CF','A2MP1','AACS'),pattern_ref = 'feature')

(k1 + k2) |>
  EMP_WGCNA_cor_analysis(method='pearson') |>
  EMP_heatmap_plot(palette = 'Spectral') 
}
}
