% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.EMP.utilis.R
\name{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis,EMPT-method}
\alias{EMP_WGCNA_cor_analysis,EMP-method}
\title{Title}
\usage{
EMP_WGCNA_cor_analysis(obj, ...)

\S4method{EMP_WGCNA_cor_analysis}{EMPT}(obj, ...)

\S4method{EMP_WGCNA_cor_analysis}{EMP}(obj, ...)
}
\arguments{
\item{obj}{x}

\item{...}{...}

\item{EMPT}{x}

\item{EMP}{x}
}
\value{
xx object

xx object

xx object
}
\description{
Title

Title

Title
}
\examples{
#x
#x
#x
}
