% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.EMP.utilis.R
\name{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis}
\alias{EMP_WGCNA_cor_analysis,EMPT-method}
\alias{EMP_WGCNA_cor_analysis,EMP-method}
\title{EMP_WGCNA_cor_analysis}
\usage{
EMP_WGCNA_cor_analysis(obj, ...)

\S4method{EMP_WGCNA_cor_analysis}{EMPT}(obj, method, coldata_to_assay, ...)

\S4method{EMP_WGCNA_cor_analysis}{EMP}(obj, select, method, ...)
}
\arguments{
\item{obj}{EMPT or MultiAssayExperiment object.}

\item{...}{Further parameters passed to the function agricolae::correlation}

\item{method}{A character string. Methods include pearson (default), spearman and kendall.}

\item{coldata_to_assay}{A series of character strings. Select the column from coldata to caculate.}

\item{select}{A character string. The experiment name in the EMP object.}
}
\value{
EMPT object
}
\description{
EMP_WGCNA_cor_analysis
}
\examples{
#x
}
