% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{humann_function_import}
\alias{humann_function_import}
\title{Title}
\usage{
humann_function_import(file = NULL, data = NULL, type)
}
\arguments{
\item{file}{A file path.}

\item{data}{A dataframe.The row must be the feature and the column is the sample.}

\item{type}{A character string. Methods include ko and ec.}
}
\value{
SummmariseExperiment object
}
\description{
Title
}
\examples{
# add example
}
