% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMP_import.R
\name{humann_taxonomy_import}
\alias{humann_taxonomy_import}
\title{Title}
\usage{
humann_taxonomy_import(x, sep = "|")
}
\arguments{
\item{x}{wait_for_add}

\item{sep}{wait_for_add}
}
\value{
xx object
}
\description{
Title
}
\examples{
# add example
}
