% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_AllClasses.R
\docType{class}
\name{multiEnrichResult-class}
\alias{multiEnrichResult-class}
\alias{show,multiEnrichResult-method}
\alias{summary,multiEnrichResult-method}
\title{Class "multiEnrichResult"
This class represents the result of multi-omics ORA enrichment analysis.}
\description{
Class "multiEnrichResult"
This class represents the result of multi-omics ORA enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{multi-omics enrichment analysis.}

\item{\code{geneID}}{enriched genes.}

\item{\code{Count}}{number of enriched genes.}

\item{\code{GeneRatio}}{Ratio of the number of genes enriched 
to the number of genes input.}

\item{\code{pvalueCutoff}}{pvalueCutoff.}

\item{\code{pAdjustMethod}}{pvalue adjust method.}

\item{\code{qvalueCutoff}}{qvalueCutoff.}

\item{\code{organism}}{organism.}

\item{\code{ontology}}{biological ontology.}

\item{\code{gene}}{input gene IDs.}

\item{\code{keytype}}{Gene ID type.}

\item{\code{universe}}{background gene.}

\item{\code{gene2Symbol}}{mapping gene to Symbol.}

\item{\code{geneSets}}{gene sets.}

\item{\code{readable}}{logical flag of gene ID in symbol or not.}

\item{\code{termsim}}{Similarity between term.}

\item{\code{method}}{method of calculating the similarity between nodes.}
}}

\seealso{
\code{\link{multiEnrichment}}
}
\author{
Erqiang Hu
}
\keyword{classes}
