% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiEnrichment_AllClasses.R
\docType{class}
\name{multiGseaResult-class}
\alias{multiGseaResult-class}
\alias{gseahResult-class}
\alias{show,multiGseaResult-method}
\alias{summary,multiGseaResult-method}
\title{Class "multiGseaResult"
This class represents the result of multi-omics GSEA analysis}
\description{
Class "multiGseaResult"
This class represents the result of multi-omics GSEA analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{multi-omics GSEA anaysis.}

\item{\code{enrichmentScore}}{enrichmentScore.}

\item{\code{NES}}{Enrichment score after normalization.}

\item{\code{rank}}{rank}

\item{\code{leading_edge}}{leading_edge.}

\item{\code{core_enrichment}}{core enrichment genes.}

\item{\code{organism}}{organism.}

\item{\code{setType}}{setType.}

\item{\code{geneSets}}{geneSets.}

\item{\code{geneList}}{order rank geneList.}

\item{\code{keytype}}{ID type of gene.}

\item{\code{permScores}}{permutation scores.}

\item{\code{params}}{parameters.}

\item{\code{gene2Symbol}}{gene ID to Symbol.}

\item{\code{readable}}{whether convert gene ID to symbol.}
}}

\author{
Erqiang Hu
}
\keyword{classes}
